Refacotr and cleanup
