+++ StationSearchCore.Domain/StationSearchCore.Domain.csproj
@@ -4,8 +4,4 @@
     <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\StationSearchCore.Domain\StationSearchCore.Domain.csproj" />
-  </ItemGroup>
-
 </Project>
