+++ StationSearchCore.Service/StationReadonlyRepository.cs
@@ -1,0 +1,2632 @@
+﻿using StationSearchCore.Domain.Interfaces;
+using System.Collections.Generic;
+
+namespace StationSearchCore.Service
+{
+    public class StationRepository : IStationRepository
+    {
+        public IList<string> GetAll() => AllStations;
+
+        private readonly IList<string> AllStations = new List<string>
+        {
+            "Aberdare Rail Station",
+            "Aberdour Rail Station",
+            "Aberdovey Rail Station",
+            "Aber Rail Station",
+            "Abergele & Pensarn Rail Station",
+            "Abercynon Rail Station",
+            "Aberdeen Rail Station",
+            "Abererch Rail Station",
+            "Abergavenny Rail Station",
+            "Aberystwyth Rail Station",
+            "Abbey Wood (London) Rail Station",
+            "Acton Bridge Rail Station",
+            "Achanalt Rail Station",
+            "Achnashellach Rail Station",
+            "Auchinleck Rail Station",
+            "Achnasheen Rail Station",
+            "Acklington Rail Station",
+            "Acle Rail Station",
+            "Acocks Green Rail Station",
+            "Accrington Rail Station",
+            "Acton Central Rail Station",
+            "Acton Main Line Rail Station",
+            "Adderley Park Rail Station",
+            "Addiewell Rail Station",
+            "Adisham Rail Station",
+            "Adlington (Cheshire) Rail Station",
+            "Addlestone Rail Station",
+            "Adlington (Lancs) Rail Station",
+            "Adwick Rail Station",
+            "Aughton Park Rail Station",
+            "Aigburth Rail Station",
+            "Ainsdale Rail Station",
+            "Aintree Rail Station",
+            "Airbles Rail Station",
+            "Airdrie Rail Station",
+            "Newcastle Airport Metro",
+            "Albany Park Rail Station",
+            "Albrighton Rail Station",
+            "Alderley Edge Rail Station",
+            "Aldermaston Rail Station",
+            "Aldershot Rail Station",
+            "Aldrington Rail Station",
+            "Liverpool South Parkway Rail Station",
+            "Alexandra Palace Rail Station",
+            "Alfreton Rail Station",
+            "Allens West Rail Station",
+            "Alloa Rail Station",
+            "Alness Rail Station",
+            "Alnmouth Rail Station",
+            "Alresford (Essex) Rail Station",
+            "Alsager Rail Station",
+            "Aylesham Rail Station",
+            "Althorne Rail Station",
+            "Altnabreac Rail Station",
+            "Alton Rail Station",
+            "Althorpe Rail Station",
+            "Altrincham Rail Station",
+            "Alvechurch Rail Station",
+            "Alexandria Rail Station",
+            "Alexandra Parade Rail Station",
+            "Ambergate Rail Station",
+            "Amersham Rail Station",
+            "Amberley Rail Station",
+            "Ammanford Rail Station",
+            "Ancaster Rail Station",
+            "Andover Rail Station",
+            "Anderston Rail Station",
+            "Anerley Rail Station",
+            "Angel Road Rail Station",
+            "Angmering Rail Station",
+            "Annan Rail Station",
+            "Ansdell & Fairhaven Rail Station",
+            "Anniesland Rail Station",
+            "Appleby Rail Station",
+            "Appledore (Kent) Rail Station",
+            "Apperley Bridge Rail Station",
+            "Appley Bridge Rail Station",
+            "Appleford Rail Station",
+            "Apsley Rail Station",
+            "Arbroath Rail Station",
+            "Arrochar & Tarbet Rail Station",
+            "Ardgay Rail Station",
+            "Ardlui Rail Station",
+            "Ardrossan Harbour Rail Station",
+            "Ardrossan South Beach Rail Station",
+            "Ardrossan Town Rail Station",
+            "Ardwick Rail Station",
+            "Argyle Street Rail Station",
+            "Arisaig Rail Station",
+            "Arlesey Rail Station",
+            "Armadale (W Lothian) Rail Station",
+            "Armathwaite Rail Station",
+            "Arnside Rail Station",
+            "Arram Rail Station",
+            "Arundel Rail Station",
+            "Ascot Rail Station",
+            "Ascott-under-Wychwood Rail Station",
+            "Ashford (Surrey) Rail Station",
+            "Ashfield Rail Station",
+            "Ashburys Rail Station",
+            "Ashchurch for Tewkesbury Rail Station",
+            "Ashtead Rail Station",
+            "Ashford International Rail Station",
+            "Ash Rail Station",
+            "Ashley Rail Station",
+            "Ashton-under-Lyne Rail Station",
+            "Ashurst (Kent) Rail Station",
+            "Ash Vale Rail Station",
+            "Ashwell & Morden Rail Station",
+            "Askam Rail Station",
+            "Aslockton Rail Station",
+            "Aspley Guise Rail Station",
+            "Aspatria Rail Station",
+            "Aston Rail Station",
+            "Attadale Rail Station",
+            "Atherton Rail Station",
+            "Atherstone Rail Station",
+            "Attenborough Rail Station",
+            "Attleborough Rail Station",
+            "Audley End Rail Station",
+            "Aviemore Rail Station",
+            "Avoncliff Rail Station",
+            "Avonmouth Rail Station",
+            "Axminster Rail Station",
+            "Aylesford Rail Station",
+            "Aylesbury Rail Station",
+            "Aylesbury Vale Parkway Rail Station",
+            "Ayr Rail Station",
+            "Bache Rail Station",
+            "Burnage Rail Station",
+            "Bagshot Rail Station",
+            "Baildon Rail Station",
+            "Balcombe Rail Station",
+            "Baldock Rail Station",
+            "Balham Rail Station",
+            "Baillieston Rail Station",
+            "Balloch Rail Station",
+            "Balmossie Rail Station",
+            "Bamford Rail Station",
+            "Banavie Rail Station",
+            "Bangor (Gwynedd) Rail Station",
+            "Banstead Rail Station",
+            "Barassie Rail Station",
+            "Bare Lane Rail Station",
+            "Bargoed Rail Station",
+            "Barking Rail Station",
+            "Barlaston Rail Station",
+            "Barming Rail Station",
+            "Barnes Rail Station",
+            "Barrow-in-Furness Rail Station",
+            "Barrow upon Soar Rail Station",
+            "Barry Rail Station",
+            "Barry Docks Rail Station",
+            "Barry Island Rail Station",
+            "Barrow Haven Rail Station",
+            "Basildon Rail Station",
+            "Bat & Ball Rail Station",
+            "Bath Spa Rail Station",
+            "Batley Rail Station",
+            "Battersea Park Rail Station",
+            "Battle Rail Station",
+            "Bayford Rail Station",
+            "Birchgrove Rail Station",
+            "Beckenham Hill Rail Station",
+            "Beckenham Junction Rail Station",
+            "Beckenham Junction Rail Station",
+            "Blackwater Rail Station",
+            "Buckley Rail Station",
+            "Birchington-on-Sea Rail Station",
+            "Beaconsfield Rail Station",
+            "Bicester North Rail Station",
+            "Bicester Village Rail Station",
+            "Bricket Wood Rail Station",
+            "Bedhampton Rail Station",
+            "Bedminster Rail Station",
+            "Bidston Rail Station",
+            "Bearley Rail Station",
+            "Beasdale Rail Station",
+            "Beauly Rail Station",
+            "Bebington Rail Station",
+            "Beccles Rail Station",
+            "Bedford Rail Station",
+            "Bedworth Rail Station",
+            "Bedwyn Rail Station",
+            "Bekesbourne Rail Station",
+            "Belmont Rail Station",
+            "Bellingham Rail Station",
+            "Belper Rail Station",
+            "Belvedere Rail Station",
+            "Bempton Rail Station",
+            "Benfleet Rail Station",
+            "Bere Alston Rail Station",
+            "Bere Ferrers Rail Station",
+            "Berkhamsted Rail Station",
+            "Bermuda Park Rail Station",
+            "Berry Brow Rail Station",
+            "Berwick (Sussex) Rail Station",
+            "Bescar Lane Rail Station",
+            "Bedford St Johns Rail Station",
+            "Beeston Rail Station",
+            "Betws-y-Coed Rail Station",
+            "Beverley Rail Station",
+            "Bexhill Rail Station",
+            "Bexley Rail Station",
+            "Baglan Rail Station",
+            "Bridge of Allan Rail Station",
+            "Barnt Green Rail Station",
+            "Birmingham International Rail Station",
+            "Jewellery Quarter Rail Station",
+            "Birmingham Moor Street Rail Station",
+            "Birmingham New Street Rail Station",
+            "Birmingham Snow Hill Rail Station",
+            "Bush Hill Park Rail Station",
+            "Bickley Rail Station",
+            "Biggleswade Rail Station",
+            "Bilbrook Rail Station",
+            "Billericay Rail Station",
+            "Billingshurst Rail Station",
+            "Bingham Rail Station",
+            "Birchwood Rail Station",
+            "Birkbeck Rail Station",
+            "Bishopton Rail Station",
+            "Burntisland Rail Station",
+            "Bitterne Rail Station",
+            "Birkdale Rail Station",
+            "Brockholes Rail Station",
+            "Bucknell Rail Station",
+            "Brockenhurst Rail Station",
+            "Berkswell Rail Station",
+            "Blaenau Ffestiniog Rail Station",
+            "Blairhill Rail Station",
+            "Blakedown Rail Station",
+            "Blantyre Rail Station",
+            "Blair Atholl Rail Station",
+            "Blaydon Rail Station",
+            "Blackhorse Road Rail Station",
+            "Blackpool Pleasure Beach Rail Station",
+            "Blackpool North Rail Station",
+            "Blackpool South Rail Station",
+            "Bleasby Rail Station",
+            "London Blackfriars Rail Station",
+            "Bellgrove Rail Station",
+            "Beaulieu Road Rail Station",
+            "Blackburn Rail Station",
+            "Blackheath Rail Station",
+            "Blackridge Rail Station",
+            "Blake Street Rail Station",
+            "Belle Vue Rail Station",
+            "Blundellsands & Crosby Rail Station",
+            "Billingham Rail Station",
+            "Bloxwich North Rail Station",
+            "Bloxwich Rail Station",
+            "Blackrod Rail Station",
+            "Bellshill Rail Station",
+            "Bletchley Rail Station",
+            "Bolton-Upon-Dearne Rail Station",
+            "Beltring Rail Station",
+            "Blythe Bridge Rail Station",
+            "Bamber Bridge Rail Station",
+            "Bramhall Rail Station",
+            "Bramley (Hants) Rail Station",
+            "Banbury Rail Station",
+            "Brandon Rail Station",
+            "Bingley Rail Station",
+            "Burnham (Berks) Rail Station",
+            "Barnhill Rail Station",
+            "Bank Hall Rail Station",
+            "Ben Rhydding Rail Station",
+            "Barnes Bridge Rail Station",
+            "Burneside (Cumbria) Rail Station",
+            "Barnsley Rail Station",
+            "Barnetby Rail Station",
+            "Bentley (Hants) Rail Station",
+            "Brentford Rail Station",
+            "Bentham Rail Station",
+            "Brunswick Rail Station",
+            "Bodmin Parkway Rail Station",
+            "Bodorgan Rail Station",
+            "Bognor Regis Rail Station",
+            "Bogston Rail Station",
+            "Bolton Rail Station",
+            "Bournemouth Rail Station",
+            "Bookham Rail Station",
+            "Bridge of Orchy Rail Station",
+            "Bootle (Cumbria) Rail Station",
+            "Bootle New Strand Rail Station",
+            "Bootle Oriel Road Rail Station",
+            "Bourne End Rail Station",
+            "Borth Rail Station",
+            "Borough Green & Wrotham Rail Station",
+            "Bosham Rail Station",
+            "Boston Rail Station",
+            "Bottesford Rail Station",
+            "Botley Rail Station",
+            "Bournville Rail Station",
+            "Bow Brickhill Rail Station",
+            "Bowes Park Rail Station",
+            "Bowling Rail Station",
+            "Box Hill & Westhumble Rail Station",
+            "Bracknell Rail Station",
+            "Bradford Forster Square Rail Station",
+            "Bradford Interchange Rail Station",
+            "Braintree Freeport Rail Station",
+            "Braintree Rail Station",
+            "Brampton (Suffolk) Rail Station",
+            "Branksome Rail Station",
+            "Broadbottom Rail Station",
+            "Brondesbury Rail Station",
+            "Brondesbury Park Rail Station",
+            "Bradford-on-Avon Rail Station",
+            "Bridgwater Rail Station",
+            "Brading Rail Station",
+            "Bridlington Rail Station",
+            "Bardon Mill Rail Station",
+            "Bordesley Rail Station",
+            "Bredbury Rail Station",
+            "Breich Rail Station",
+            "Broughty Ferry Rail Station",
+            "Broomfleet Rail Station",
+            "Bargeddie Rail Station",
+            "Bridgend Rail Station",
+            "Brigg Rail Station",
+            "London Road (Brighton) Rail Station",
+            "Brighton Rail Station",
+            "Barrhead Rail Station",
+            "Barnham Rail Station",
+            "Brighouse Rail Station",
+            "Bridgeton Rail Station",
+            "Brierfield Rail Station",
+            "Brimsdown Rail Station",
+            "Briton Ferry Rail Station",
+            "Brithdir Rail Station",
+            "Brixton Rail Station",
+            "Brookmans Park Rail Station",
+            "Conway Park Rail Station",
+            "Birkenhead Central Rail Station",
+            "Birkenhead North Rail Station",
+            "Birkenhead Park Rail Station",
+            "Brookwood Rail Station",
+            "Berrylands Rail Station",
+            "Burley Park Rail Station",
+            "Bromborough Rail Station",
+            "Bromborough Rake Rail Station",
+            "Bromley Cross Rail Station",
+            "Bramley (West Yorks) Rail Station",
+            "Barmouth Rail Station",
+            "Brampton (Cumbria) Rail Station",
+            "Bromsgrove Rail Station",
+            "Branchton Rail Station",
+            "Brinnington Rail Station",
+            "Barnehurst Rail Station",
+            "Burnham-on-Crouch Rail Station",
+            "Barnstaple Rail Station",
+            "Berney Arms Rail Station",
+            "Broad Green Rail Station",
+            "Brockley Rail Station",
+            "Bromley North Rail Station",
+            "Bromley South Rail Station",
+            "Broome Rail Station",
+            "Brora Rail Station",
+            "Brough Rail Station",
+            "Broxbourne Rail Station",
+            "Barrhill Rail Station",
+            "Barry Links Rail Station",
+            "Burscough Bridge Rail Station",
+            "Burscough Junction Rail Station",
+            "Burnside (Strathclyde) Rail Station",
+            "Broadstairs Rail Station",
+            "Bristol Temple Meads Rail Station",
+            "Bristol Parkway Rail Station",
+            "Barton-on-Humber Rail Station",
+            "Burton Joyce Rail Station",
+            "Brentwood Rail Station",
+            "Bruce Grove Rail Station",
+            "Brundall Rail Station",
+            "Brundall Gardens Rail Station",
+            "Brunstane Rail Station",
+            "Bruton Rail Station",
+            "Berwick-upon-Tweed Rail Station",
+            "Brockley Whins Rail Station",
+            "Bryn Rail Station",
+            "Braystones Rail Station",
+            "Bishop Auckland Rail Station",
+            "Bishop Auckland Rail Station",
+            "Bescot Stadium Rail Station",
+            "Bearsden Rail Station",
+            "Bishopbriggs Rail Station",
+            "Bishops Stortford Rail Station",
+            "Basingstoke Rail Station",
+            "Bishop's Lydeard",
+            "Bishopstone Rail Station",
+            "Bearsted Rail Station",
+            "Bury St Edmunds Rail Station",
+            "Betchworth Rail Station",
+            "Bathgate Rail Station",
+            "Bethnal Green Rail Station",
+            "Butlers Lane Rail Station",
+            "Battlesbridge Rail Station",
+            "Bentley (S Yorks) Rail Station",
+            "Battersby Rail Station",
+            "Buckenham Rail Station",
+            "Bugle Rail Station",
+            "Builth Road Rail Station",
+            "Bulwell Rail Station",
+            "Bures Rail Station",
+            "Burgess Hill Rail Station",
+            "Burley-in-Wharfedale Rail Station",
+            "Burnley Barracks Rail Station",
+            "Burnley Central Rail Station",
+            "Burnley Manchester Road Rail Station",
+            "Bursledon Rail Station",
+            "Burton-on-Trent Rail Station",
+            "Busby Rail Station",
+            "Bushey Rail Station",
+            "Bushey Rail Station",
+            "Buxton Rail Station",
+            "Bexleyheath Rail Station",
+            "Buxted Rail Station",
+            "Byfleet & New Haw Rail Station",
+            "Bynea Rail Station",
+            "Cadoxton Rail Station",
+            "Caergwrle Rail Station",
+            "Calstock Rail Station",
+            "Cambridge Rail Station",
+            "Camelon Rail Station",
+            "Cambridge Heath (London) Rail Station",
+            "Canley Rail Station",
+            "London Cannon Street Rail Station",
+            "Carbis Bay Rail Station",
+            "Cardiff Bay Rail Station",
+            "Cardiff Queen Street Rail Station",
+            "Carfin Rail Station",
+            "Cark Rail Station",
+            "Carlisle Rail Station",
+            "Carmyle Rail Station",
+            "Carrbridge Rail Station",
+            "Caersws Rail Station",
+            "Castleford Rail Station",
+            "Caterham Rail Station",
+            "Catford Bridge Rail Station",
+            "Catford Rail Station",
+            "Cathays Rail Station",
+            "Cattal Rail Station",
+            "Causeland Rail Station",
+            "Castle Bar Park Rail Station",
+            "Colwyn Bay Rail Station",
+            "Crowborough Rail Station",
+            "Cobham & Stoke d'Abernon Rail Station",
+            "Camborne Rail Station",
+            "Cranbrook Rail Station",
+            "Castle Cary Rail Station",
+            "Cathcart Rail Station",
+            "Caldicot Rail Station",
+            "Cardenden Rail Station",
+            "Cardonald Rail Station",
+            "Cardross Rail Station",
+            "Cefn-y-Bedd Rail Station",
+            "Croftfoot Rail Station",
+            "Crossflatts Rail Station",
+            "Crofton Park Rail Station",
+            "Chafford Hundred Rail Station",
+            "Chapelton (Devon) Rail Station",
+            "Chartham Rail Station",
+            "Chassen Road Rail Station",
+            "Chatham Rail Station",
+            "Charlbury Rail Station",
+            "Chichester Rail Station",
+            "Cheadle Hulme Rail Station",
+            "Cheddington Rail Station",
+            "Chadwell Heath Rail Station",
+            "Cheam Rail Station",
+            "Chapel-en-le-Frith Rail Station",
+            "Chelford Rail Station",
+            "Chepstow Rail Station",
+            "Cheshunt Rail Station",
+            "Chetnole Rail Station",
+            "Chesterfield Rail Station",
+            "Church Fenton Rail Station",
+            "Chathill Rail Station",
+            "Chingford Rail Station",
+            "Chippenham Rail Station",
+            "Chirk Rail Station",
+            "Chiswick Rail Station",
+            "Chalfont & Latimer Rail Station",
+            "Chalkwell Rail Station",
+            "Chelmsford Rail Station",
+            "Chelsfield Rail Station",
+            "Cheltenham Spa Rail Station",
+            "Chilworth Rail Station",
+            "Chinley Rail Station",
+            "Cholsey Rail Station",
+            "Buckshaw Parkway Rail Station",
+            "Chorley Rail Station",
+            "Church & Oswaldtwistle Rail Station",
+            "Chapeltown Rail Station",
+            "Chappel & Wakes Colne Rail Station",
+            "Christchurch Rail Station",
+            "Charing Cross (Glasgow) Rail Station",
+            "Christs Hospital Rail Station",
+            "Chorleywood Rail Station",
+            "London Charing Cross Rail Station",
+            "Chipstead Rail Station",
+            "Chislehurst Rail Station",
+            "Chester Road Rail Station",
+            "Chessington North Rail Station",
+            "Chessington South Rail Station",
+            "Chester Rail Station",
+            "Chestfield & Swalecliffe Rail Station",
+            "Chatelherault Rail Station",
+            "Cherry Tree Rail Station",
+            "Chertsey Rail Station",
+            "Cilmeri Rail Station",
+            "Corkerhill Rail Station",
+            "Clacton-on-Sea Rail Station",
+            "Clandon Rail Station",
+            "Clapton Rail Station",
+            "Clarbeston Road Rail Station",
+            "Claverdon Rail Station",
+            "Hythe (Essex) Rail Station",
+            "Colchester Town Rail Station",
+            "Colchester Rail Station",
+            "Caledonian Road & Barnsbury Rail Station",
+            "Caldercruix Rail Station",
+            "Cleland Rail Station",
+            "Clifton Down Rail Station",
+            "Collingham Rail Station",
+            "Clitheroe Rail Station",
+            "Chilham Rail Station",
+            "Clock House Rail Station",
+            "Clapham High Street Rail Station",
+            "Clapham Junction Rail Station",
+            "Clapham Junction Rail Station",
+            "Clapham Junction Rail Station",
+            "Clapham Junction Rail Station",
+            "Clapham Junction Rail Station",
+            "Clapham (N Yorks) Rail Station",
+            "Clarkston Rail Station",
+            "Chester-le-Street Rail Station",
+            "Cleethorpes Rail Station",
+            "Clifton (Manchester) Rail Station",
+            "Colwall Rail Station",
+            "Clydebank Rail Station",
+            "Claygate Rail Station",
+            "Clunderwen Rail Station",
+            "Cwmbach Rail Station",
+            "Camberley Rail Station",
+            "Cumbernauld Rail Station",
+            "Cambuslang Rail Station",
+            "Camden Road Rail Station",
+            "Cromer Rail Station",
+            "Carmarthen Rail Station",
+            "Cannock Rail Station",
+            "Canada Water Rail Station",
+            "Canonbury Rail Station",
+            "Carnoustie Rail Station",
+            "Canterbury East Rail Station",
+            "Canterbury West Rail Station",
+            "Cantley Rail Station",
+            "Carntyne Rail Station",
+            "Coatbridge Central Rail Station",
+            "Coatbridge Sunnyside Rail Station",
+            "Coatdyke Rail Station",
+            "Cooden Beach Rail Station",
+            "Codsall Rail Station",
+            "Cogan Rail Station",
+            "Cooksbridge Rail Station",
+            "Coleshill Parkway Rail Station",
+            "Colne Rail Station",
+            "Coulsdon South Rail Station",
+            "Coulsdon Town Rail Station",
+            "Combe (Oxon) Rail Station",
+            "Commondale Rail Station",
+            "Conisbrough Rail Station",
+            "Congleton Rail Station",
+            "Connel Ferry Rail Station",
+            "Conon Bridge Rail Station",
+            "Cononley Rail Station",
+            "Conwy Rail Station",
+            "Cookham Rail Station",
+            "Coombe Junction Halt (Rail Station)",
+            "Copplestone Rail Station",
+            "Corby Rail Station",
+            "Corpach Rail Station",
+            "Corrour Rail Station",
+            "Coryton Rail Station",
+            "Coseley Rail Station",
+            "Cosford Rail Station",
+            "Cosham Rail Station",
+            "Cottingley Rail Station",
+            "Coventry Arena Rail Station",
+            "Coventry Rail Station",
+            "Cowden Rail Station",
+            "Capenhurst Rail Station",
+            "Cradley Heath Rail Station",
+            "Craven Arms Rail Station",
+            "Crawley Rail Station",
+            "Corbridge Rail Station",
+            "Corkickle Rail Station",
+            "Cardiff Central Rail Station",
+            "Crediton Rail Station",
+            "Cressing Rail Station",
+            "Crewe Rail Station",
+            "Crayford Rail Station",
+            "Craigendoran Rail Station",
+            "Crews Hill Rail Station",
+            "Criccieth Rail Station",
+            "Cricklewood Rail Station",
+            "Crewkerne Rail Station",
+            "Carluke Rail Station",
+            "Charlton Rail Station",
+            "Cramlington Rail Station",
+            "Carnforth Rail Station",
+            "Charing (Kent) Rail Station",
+            "Crianlarich Rail Station",
+            "Crookston Rail Station",
+            "Cromford Rail Station",
+            "Crosshill Rail Station",
+            "Croston Rail Station",
+            "Crouch Hill Rail Station",
+            "Crowle Rail Station",
+            "Croy Rail Station",
+            "Caerphilly Rail Station",
+            "Carpenders Park Rail Station",
+            "Carshalton Beeches Rail Station",
+            "Carshalton Rail Station",
+            "Crosskeys Rail Station",
+            "Crossmyloof Rail Station",
+            "Cressington Rail Station",
+            "Carstairs Rail Station",
+            "Carlton Rail Station",
+            "Cartsdyke Rail Station",
+            "Crowhurst Rail Station",
+            "Creswell (Derbys) Rail Station",
+            "Crowthorne Rail Station",
+            "Crystal Palace Rail Station",
+            "Imperial Wharf Rail Station",
+            "Chandlers Ford Rail Station",
+            "Cross Gates Rail Station",
+            "Castleton (Manchester) Rail Station",
+            "Church Stretton Rail Station",
+            "City Thameslink Rail Station",
+            "Collington Rail Station",
+            "Castleton Moor Rail Station",
+            "Cottingham Rail Station",
+            "Cuddington Rail Station",
+            "Cuffley Rail Station",
+            "Culham Rail Station",
+            "Culrain Rail Station",
+            "Cupar Rail Station",
+            "Curriehill Rail Station",
+            "Cowdenbeath Rail Station",
+            "Cwmbran Rail Station",
+            "Cuxton Rail Station",
+            "Cynghordy Rail Station",
+            "Daisy Hill Rail Station",
+            "Dalgety Bay Rail Station",
+            "Dalmuir Rail Station",
+            "Dalry Rail Station",
+            "Dalston Junction Rail Station",
+            "Dalston Kingsland Rail Station",
+            "Dalston (Cumbria) Rail Station",
+            "Dalton Rail Station",
+            "Dalwhinnie Rail Station",
+            "Danby Rail Station",
+            "Danescourt Rail Station",
+            "Danzey Rail Station",
+            "Darnall Rail Station",
+            "Darsham Rail Station",
+            "Dartford Rail Station",
+            "Darwen Rail Station",
+            "Datchet Rail Station",
+            "Davenport Rail Station",
+            "Dawlish Rail Station",
+            "Dawlish Warren Rail Station",
+            "Kelvindale Rail Station",
+            "Deal Rail Station",
+            "Dean Rail Station",
+            "Deganwy Rail Station",
+            "Deighton Rail Station",
+            "Delamere Rail Station",
+            "Denham Rail Station",
+            "Denham Golf Club Rail Station",
+            "Denmark Hill Rail Station",
+            "Dent Rail Station",
+            "Dorking Deepdene Rail Station",
+            "Deptford Rail Station",
+            "Derby Road (Ipswich) Rail Station",
+            "Devonport Rail Station",
+            "Dingle Road Rail Station",
+            "Dagenham Dock Rail Station",
+            "Didcot Parkway Rail Station",
+            "Digby & Sowton Rail Station",
+            "Dilton Marsh Rail Station",
+            "Dinas Powys Rail Station",
+            "Dinas (Rhondda) Rail Station",
+            "Dingwall Rail Station",
+            "Dinsdale Rail Station",
+            "Disley Rail Station",
+            "Diss Rail Station",
+            "Dalmally Rail Station",
+            "Dalmarnock Rail Station",
+            "Dalmeny Rail Station",
+            "Dalreoch Rail Station",
+            "Darlington Rail Station",
+            "Dolwyddelan Rail Station",
+            "Dumbarton Central Rail Station",
+            "Dumbarton East Rail Station",
+            "Denby Dale Rail Station",
+            "Dunfermline Town Rail Station",
+            "Dunfermline Queen Margaret Rail Station",
+            "Dunlop Rail Station",
+            "Dunrobin Castle Rail Station",
+            "Dunston Rail Station",
+            "Dinting Rail Station",
+            "Denton Rail Station",
+            "Dockyard (Plymouth) Rail Station",
+            "Dodworth Rail Station",
+            "Dolau Rail Station",
+            "Doleham Rail Station",
+            "Dolgarrog Rail Station",
+            "Doncaster Rail Station",
+            "Dore & Totley Rail Station",
+            "Dorridge Rail Station",
+            "Dorking Rail Station",
+            "Dormans Rail Station",
+            "Dover Priory Rail Station",
+            "Dovercourt Rail Station",
+            "Dovey Junction Rail Station",
+            "Downham Market Rail Station",
+            "Drayton Green Rail Station",
+            "Derby Rail Station",
+            "Dorchester South Rail Station",
+            "Dorchester West Rail Station",
+            "Drem Rail Station",
+            "Durham Rail Station",
+            "Driffield Rail Station",
+            "Drigg Rail Station",
+            "Dorking West Rail Station",
+            "Drumchapel Rail Station",
+            "Drumfrochar Rail Station",
+            "Drumgelloch Rail Station",
+            "Durrington-on-Sea Rail Station",
+            "Dronfield Rail Station",
+            "Darton Rail Station",
+            "Droitwich Spa Rail Station",
+            "Drumry Rail Station",
+            "Drayton Park Rail Station",
+            "Dunton Green Rail Station",
+            "Duddeston Rail Station",
+            "Dudley Port Rail Station",
+            "Duffield Rail Station",
+            "Duirinish Rail Station",
+            "Duke Street Rail Station",
+            "Dullingham Rail Station",
+            "Dumbreck Rail Station",
+            "Dumfries Rail Station",
+            "Dumpton Park Rail Station",
+            "Dunblane Rail Station",
+            "Dunbar Rail Station",
+            "Duncraig Rail Station",
+            "Dundee Rail Station",
+            "Dunkeld & Birnam Rail Station",
+            "Cam & Dursley Rail Station",
+            "Dove Holes Rail Station",
+            "Dewsbury Rail Station",
+            "Dyce Rail Station",
+            "Dyffryn Ardudwy Rail Station",
+            "Ealing Broadway Rail Station",
+            "Earlswood (Surrey) Rail Station",
+            "Eastbrook Rail Station",
+            "East Boldon Rail Station",
+            "Eastbourne Rail Station",
+            "Ebbsfleet International Rail Station",
+            "Ebbw Vale Parkway Rail Station",
+            "Ebbw Vale Town Rail Station",
+            "Eccles Rail Station",
+            "Eccles Road Rail Station",
+            "Eccleston Park Rail Station",
+            "East Croydon Rail Station",
+            "Edale Rail Station",
+            "Edge Hill Rail Station",
+            "East Didsbury Rail Station",
+            "Edinburgh Rail Station",
+            "Edinburgh Park Rail Station",
+            "Edmonton Green Rail Station",
+            "Edenbridge Rail Station",
+            "Edenbridge Town Rail Station",
+            "Eden Park Rail Station",
+            "East Dulwich Rail Station",
+            "East Farleigh Rail Station",
+            "Effingham Junction Rail Station",
+            "Eggesford Rail Station",
+            "Egham Rail Station",
+            "Eaglescliffe Rail Station",
+            "East Garforth Rail Station",
+            "East Grinstead Rail Station",
+            "Egton Rail Station",
+            "East Kilbride Rail Station",
+            "Eastleigh Rail Station",
+            "Elgin Rail Station",
+            "Elmers End Rail Station",
+            "Elmswell Rail Station",
+            "Elmstead Woods Rail Station",
+            "Elephant & Castle Rail Station",
+            "Elsecar Rail Station",
+            "Elsenham Rail Station",
+            "Ellesmere Port Rail Station",
+            "Eltham Rail Station",
+            "Elton & Orston Rail Station",
+            "Elstree & Borehamwood Rail Station",
+            "Earlswood (West Midlands) Rail Station",
+            "Ely Rail Station",
+            "East Malling Rail Station",
+            "East Midlands Parkway Rail Station",
+            "Emerson Park Rail Station",
+            "Emsworth Rail Station",
+            "Enfield Chase Rail Station",
+            "Enfield Lock Rail Station",
+            "Enfield Town Rail Station",
+            "Entwistle Rail Station",
+            "Epsom Downs Rail Station",
+            "Epsom Rail Station",
+            "Erdington Rail Station",
+            "Energlyn & Churchill Park Rail Station",
+            "Eridge Rail Station",
+            "Erith Rail Station",
+            "Earlsfield Rail Station",
+            "Earlestown Rail Station",
+            "Earley Rail Station",
+            "Esher Rail Station",
+            "Easterhouse Rail Station",
+            "Eskbank Rail Station",
+            "Essex Road Rail Station",
+            "Mitcham Eastfields Rail Station",
+            "Eastham Rake Rail Station",
+            "Eastrington Rail Station",
+            "Etchingham Rail Station",
+            "East Tilbury Rail Station",
+            "London Euston Rail Station",
+            "Euxton Balshaw Lane Rail Station",
+            "Evesham Rail Station",
+            "Ewell East Rail Station",
+            "Ewell West Rail Station",
+            "East Worthing Rail Station",
+            "Exeter Central Rail Station",
+            "Exeter St Davids Rail Station",
+            "Exeter St Thomas Rail Station",
+            "Exhibition Centre (Glasgow) Rail Station",
+            "Exmouth Rail Station",
+            "Exton Rail Station",
+            "Eynsford Rail Station",
+            "Fairlie Rail Station",
+            "Falconwood Rail Station",
+            "Falkirk Grahamston Rail Station",
+            "Falkirk High Rail Station",
+            "Falmouth Docks Rail Station",
+            "Falmouth Town Rail Station",
+            "North Fambridge Rail Station",
+            "Falmer Rail Station",
+            "Fareham Rail Station",
+            "Farnham Rail Station",
+            "Fauldhouse Rail Station",
+            "Faversham Rail Station",
+            "Faygate Rail Station",
+            "Fazakerley Rail Station",
+            "Fearn Rail Station",
+            "Feltham Rail Station",
+            "London Fenchurch Street Rail Station",
+            "Feniton Rail Station",
+            "Fernhill Rail Station",
+            "Ferriby Rail Station",
+            "Ffairfach Rail Station",
+            "Fishguard Harbour Rail Station",
+            "Fishguard & Goodwick Rail Station",
+            "Filey Rail Station",
+            "Filton Abbey Wood Rail Station",
+            "Finstock Rail Station",
+            "Fiskerton Rail Station",
+            "Fitzwilliam Rail Station",
+            "Five Ways Rail Station",
+            "Fleet Rail Station",
+            "Flimby Rail Station",
+            "Flint Rail Station",
+            "Flitwick Rail Station",
+            "Flixton Rail Station",
+            "Folkestone Central Rail Station",
+            "Folkestone West Rail Station",
+            "Falls of Cruachan Rail Station",
+            "Flowery Field Rail Station",
+            "Felixstowe Rail Station",
+            "Finchley Road & Frognal Rail Station",
+            "Finsbury Park Rail Station",
+            "Ford Rail Station",
+            "Forest Hill Rail Station",
+            "Formby Rail Station",
+            "Forres Rail Station",
+            "Forsinard Rail Station",
+            "Four Oaks Rail Station",
+            "Foxfield Rail Station",
+            "Foxton Rail Station",
+            "Frant Rail Station",
+            "Fratton Rail Station",
+            "Fairbourne Rail Station",
+            "Farnborough (Main) Rail Station",
+            "Farnborough North Rail Station",
+            "Frodsham Rail Station",
+            "Freshford Rail Station",
+            "Fairfield Rail Station",
+            "Frimley Rail Station",
+            "Frinton-on-Sea Rail Station",
+            "Farncombe Rail Station",
+            "Farringdon (London) Rail Station",
+            "Farningham Road Rail Station",
+            "Farnworth Rail Station",
+            "Frome Rail Station",
+            "Frosterley Rail Station",
+            "Freshfield Rail Station",
+            "Forest Gate Rail Station",
+            "Fort Matilda Rail Station",
+            "Fort William Rail Station",
+            "Fairwater Rail Station",
+            "Frizinghall Rail Station",
+            "Fishbourne Rail Station",
+            "Fishersgate Rail Station",
+            "Fenny Stratford Rail Station",
+            "Featherstone Rail Station",
+            "Fulwell Rail Station",
+            "Furness Vale Rail Station",
+            "Furze Platt Rail Station",
+            "Ferryside Rail Station",
+            "Galashiels Rail Station",
+            "Smethwick Galton Bridge Rail Station",
+            "Smethwick Galton Bridge High Level Rail Station",
+            "Garforth Rail Station",
+            "Garrowhill Rail Station",
+            "Garsdale Rail Station",
+            "Garth (Powys) Rail Station",
+            "Garve Rail Station",
+            "Garswood Rail Station",
+            "Gathurst Rail Station",
+            "Gatley Rail Station",
+            "Gainsborough Lea Road Rail Station",
+            "Greenbank Rail Station",
+            "Gainsborough Central Rail Station",
+            "Godalming Rail Station",
+            "Gerrards Cross Rail Station",
+            "Greenfield Rail Station",
+            "Greenford Rail Station",
+            "Giggleswick Rail Station",
+            "Guide Bridge Rail Station",
+            "Gidea Park Rail Station",
+            "Giffnock Rail Station",
+            "Gipsy Hill Rail Station",
+            "Girvan Rail Station",
+            "Glaisdale Rail Station",
+            "Glasshoughton Rail Station",
+            "Glazebrook Rail Station",
+            "Gilberdyke Rail Station",
+            "Goldthorpe Rail Station",
+            "Gilfach Fargoed Rail Station",
+            "Glasgow Central Rail Station",
+            "Glasgow Central Low Level Rail Station",
+            "High Street (Glasgow) Rail Station",
+            "Glengarnock Rail Station",
+            "Glasgow Queen Street Rail Station",
+            "Glasgow Queen Street Low Level Rail Station",
+            "Gillingham (Dorset) Rail Station",
+            "Glan Conwy Rail Station",
+            "Gleneagles Rail Station",
+            "Glenfinnan Rail Station",
+            "Gillingham (Kent) Rail Station",
+            "Glenrothes with Thornton Rail Station",
+            "Gloucester Rail Station",
+            "Glossop Rail Station",
+            "Glynde Rail Station",
+            "Greenhithe for Bluewater Rail Station",
+            "Green Lane Rail Station",
+            "Gunnersbury Rail Station",
+            "Grangetown (Cardiff) Rail Station",
+            "Greenwich Rail Station",
+            "Gobowen Rail Station",
+            "Godley Rail Station",
+            "Goodmayes Rail Station",
+            "Godstone Rail Station",
+            "Golf Street Rail Station",
+            "Golspie Rail Station",
+            "Gomshall Rail Station",
+            "Goole Rail Station",
+            "Goostrey Rail Station",
+            "Goring & Streatley Rail Station",
+            "Gordon Hill Rail Station",
+            "Gorebridge Rail Station",
+            "Goring-by-Sea Rail Station",
+            "Gorton Rail Station",
+            "Gospel Oak Rail Station",
+            "Gourock Rail Station",
+            "Grange-over-Sands Rail Station",
+            "Gowerton Rail Station",
+            "Goxhill Rail Station",
+            "Grateley Rail Station",
+            "Gravelly Hill Rail Station",
+            "Grays Rail Station",
+            "Green Road Rail Station",
+            "Greenock Central Rail Station",
+            "Greenock West Rail Station",
+            "Gretna Green Rail Station",
+            "Georgemas Junction Rail Station",
+            "Gargrave Rail Station",
+            "Garelochhead Rail Station",
+            "Grimsby Docks Rail Station",
+            "Grimsby Town Rail Station",
+            "Grindleford Rail Station",
+            "Greenfaulds Rail Station",
+            "Grosmont Rail Station",
+            "Grange Park Rail Station",
+            "Garston (Herts) Rail Station",
+            "Gartcosh Rail Station",
+            "Garth (Bridgend) Rail Station",
+            "Grove Park Rail Station",
+            "Gravesend Rail Station",
+            "Garscadden Rail Station",
+            "Great Ayton Rail Station",
+            "Great Bentley Rail Station",
+            "Great Chesterford Rail Station",
+            "Great Coates Rail Station",
+            "Grantham Rail Station",
+            "Great Malvern Rail Station",
+            "Great Missenden Rail Station",
+            "Metrocentre Rail Station",
+            "Gatwick Airport Rail Station",
+            "Guildford Rail Station",
+            "Guiseley Rail Station",
+            "Gunnislake Rail Station",
+            "Gunton Rail Station",
+            "Gwersyllt Rail Station",
+            "Gypsy Lane Rail Station",
+            "Habrough Rail Station",
+            "Hackney Central Rail Station",
+            "Hackney Wick Rail Station",
+            "Hadfield Rail Station",
+            "Haddiscoe Rail Station",
+            "Hadley Wood Rail Station",
+            "Haddenham & Thame Parkway Rail Station",
+            "Hag Fold Rail Station",
+            "Haggerston Rail Station",
+            "Hagley Rail Station",
+            "Harrogate Rail Station",
+            "Hackney Downs Rail Station",
+            "Hale Rail Station",
+            "Halling Rail Station",
+            "Hall Road Rail Station",
+            "Halesworth Rail Station",
+            "Halewood Rail Station",
+            "Hammerton Rail Station",
+            "Hamilton Central Rail Station",
+            "Hamilton West Rail Station",
+            "Hampton (London) Rail Station",
+            "Hamstead (Birmingham) Rail Station",
+            "Birkenhead Hamilton Square Rail Station",
+            "Hampton Wick Rail Station",
+            "Hanborough Rail Station",
+            "Hanwell Rail Station",
+            "Hapton Rail Station",
+            "Harlesden Rail Station",
+            "Hairmyres Rail Station",
+            "Harrow-on-the-Hill Rail Station",
+            "Hartford Rail Station",
+            "Harwich Town Rail Station",
+            "Haslemere Rail Station",
+            "Hassocks Rail Station",
+            "Hastings Rail Station",
+            "Hatfield (Herts) Rail Station",
+            "Hathersage Rail Station",
+            "Hattersley Rail Station",
+            "Hatton (Warks) Rail Station",
+            "Havant Rail Station",
+            "Havenhouse Rail Station",
+            "Haverfordwest Rail Station",
+            "Hawarden Rail Station",
+            "Hawarden Bridge Rail Station",
+            "Hayes & Harlington Rail Station",
+            "Hayle Rail Station",
+            "Haymarket Rail Station",
+            "Hayes (Kent) Rail Station",
+            "Hazel Grove Rail Station",
+            "Hebden Bridge Rail Station",
+            "Hollingbourne Rail Station",
+            "High Brooms Rail Station",
+            "Hubberts Bridge Rail Station",
+            "Hartlebury Rail Station",
+            "Holmes Chapel Rail Station",
+            "Heckington Rail Station",
+            "Hampton Court Rail Station",
+            "Hednesford Rail Station",
+            "Hendon Rail Station",
+            "Huddersfield Rail Station",
+            "Headcorn Rail Station",
+            "Healing Rail Station",
+            "Heath High Level Rail Station",
+            "Heath Low Level Rail Station",
+            "Hedge End Rail Station",
+            "Headingley Rail Station",
+            "Headstone Lane Rail Station",
+            "Hellifield Rail Station",
+            "Helmsdale Rail Station",
+            "Helsby Rail Station",
+            "Hemel Hempstead Rail Station",
+            "Hengoed Rail Station",
+            "Henley-in-Arden Rail Station",
+            "Hensall Rail Station",
+            "Hereford Rail Station",
+            "Herne Hill Rail Station",
+            "Hersham Rail Station",
+            "Hertford East Rail Station",
+            "Hessle Rail Station",
+            "Heswall Rail Station",
+            "Hever Rail Station",
+            "Heworth Rail Station",
+            "Hexham Rail Station",
+            "Heyford Rail Station",
+            "Heysham Port Rail Station",
+            "Hertford North Rail Station",
+            "Hatfield Peverel Rail Station",
+            "Highbridge & Burnham-on-Sea Rail Station",
+            "Highbury & Islington Rail Station",
+            "Highams Park Rail Station",
+            "Heighington Rail Station",
+            "Highbury & Islington Rail Station",
+            "Higham Rail Station",
+            "Hillside Rail Station",
+            "Hilsea Rail Station",
+            "Hinckley Rail Station",
+            "Hindley Rail Station",
+            "Hinton Admiral Rail Station",
+            "Hitchin Rail Station",
+            "Hightown Rail Station",
+            "Hall i' th' Wood Rail Station",
+            "Hackbridge Rail Station",
+            "Heald Green Rail Station",
+            "Hildenborough Rail Station",
+            "Hillfoot Rail Station",
+            "Halifax Rail Station",
+            "Hall Green Rail Station",
+            "Holmwood Rail Station",
+            "Hillington East Rail Station",
+            "Hillington West Rail Station",
+            "Helensburgh Central Rail Station",
+            "Helensburgh Upper Rail Station",
+            "Haltwhistle Rail Station",
+            "Holyhead Rail Station",
+            "Hamble Rail Station",
+            "Hampden Park (Sussex) Rail Station",
+            "Humphrey Park Rail Station",
+            "Hampstead Heath Rail Station",
+            "Hampton-in-Arden Rail Station",
+            "Ham Street Rail Station",
+            "Hamworthy Rail Station",
+            "Hinchley Wood Rail Station",
+            "Handforth Rail Station",
+            "The Hawthorns Rail Station",
+            "Harrington Rail Station",
+            "Henley-on-Thames Rail Station",
+            "Huntly Rail Station",
+            "Huntingdon Rail Station",
+            "Hockley Rail Station",
+            "Holton Heath Rail Station",
+            "Holytown Rail Station",
+            "Homerton Rail Station",
+            "Honiton Rail Station",
+            "Honley Rail Station",
+            "Honor Oak Park Rail Station",
+            "Honeybourne Rail Station",
+            "Hook Rail Station",
+            "Hooton Rail Station",
+            "Hope (Flintshire) Rail Station",
+            "Hope (Derbyshire) Rail Station",
+            "Hopton Heath Rail Station",
+            "Horley Rail Station",
+            "Horsham Rail Station",
+            "Horwich Parkway Rail Station",
+            "Hoscar Rail Station",
+            "Hough Green Rail Station",
+            "Hounslow Rail Station",
+            "Hove Rail Station",
+            "Howden Rail Station",
+            "Howwood (Renfrewshire) Rail Station",
+            "How Wood (Herts) Rail Station",
+            "Hoxton Rail Station",
+            "Hoylake Rail Station",
+            "Harringay Rail Station",
+            "Harringay Green Lanes Rail Station",
+            "Harlech Rail Station",
+            "Harold Wood Rail Station",
+            "Harlington (Beds) Rail Station",
+            "Harling Road Rail Station",
+            "Harlow Mill Rail Station",
+            "Harlow Town Rail Station",
+            "Hornbeam Park Rail Station",
+            "Herne Bay Rail Station",
+            "Hornsey Rail Station",
+            "Harrow & Wealdstone Rail Station",
+            "Harrow & Wealdstone Rail Station",
+            "Harpenden Rail Station",
+            "Horsley Rail Station",
+            "Harrietsham Rail Station",
+            "Hurst Green Rail Station",
+            "Hartlepool Rail Station",
+            "Hartwood Rail Station",
+            "Horsforth Rail Station",
+            "Hatch End Rail Station",
+            "Heaton Chapel Rail Station",
+            "Hatfield & Stainforth Rail Station",
+            "Hither Green Rail Station",
+            "Hutton Cranswick Rail Station",
+            "Horton-in-Ribblesdale Rail Station",
+            "Heathrow Airport Terminal 5",
+            "Heathrow Terminals 1-3 Rail Station",
+            "Heathrow Airport Central Bus Stn (Rail-Air)",
+            "Heathrow Terminal 4 (Rail-Air)",
+            "Heathrow Terminal 4 Rail Station",
+            "Heathrow Terminal 5 Rail Station",
+            "Hucknall Rail Station",
+            "Hull Rail Station",
+            "Huncoat Rail Station",
+            "Hungerford Rail Station",
+            "Hunmanby Rail Station",
+            "Hunts Cross Rail Station",
+            "Huyton Rail Station",
+            "Hawkhead Rail Station",
+            "High Wycombe Rail Station",
+            "Haydon Bridge Rail Station",
+            "Hyde Central Rail Station",
+            "Hyde North Rail Station",
+            "Haydons Road Rail Station",
+            "Hykeham Rail Station",
+            "Hyndland Rail Station",
+            "Haywards Heath Rail Station",
+            "IBM (Greenock) Rail Station",
+            "Ifield Rail Station",
+            "Ilford Rail Station",
+            "Ilkeston Rail Station",
+            "Ilkley Rail Station",
+            "Ince (Manchester) Rail Station",
+            "Ince & Elton Rail Station",
+            "Ingatestone Rail Station",
+            "Insch Rail Station",
+            "Inverkip Rail Station",
+            "Invergowrie Rail Station",
+            "Invershin Rail Station",
+            "Inverurie Rail Station",
+            "Ipswich Rail Station",
+            "Irlam Rail Station",
+            "Irvine Rail Station",
+            "Isleworth Rail Station",
+            "Islip Rail Station",
+            "Iver Rail Station",
+            "Invergordon Rail Station",
+            "Inverkeithing Rail Station",
+            "Inverness Rail Station",
+            "Ivybridge Rail Station",
+            "James Cook University Hospital Rail Station",
+            "Liverpool James Street Rail Station",
+            "Johnston (Pembrokeshire) Rail Station",
+            "Johnstone Rail Station",
+            "Jordanhill Rail Station",
+            "Kirkcaldy Rail Station",
+            "Kidderminster Rail Station",
+            "Kearsley Rail Station",
+            "Keighley Rail Station",
+            "Keith Rail Station",
+            "Kelvedon Rail Station",
+            "Kemble Rail Station",
+            "Kemsing Rail Station",
+            "Kendal Rail Station",
+            "Kennishead Rail Station",
+            "Kennett Rail Station",
+            "Kensington (Olympia) Rail Station",
+            "Kensal Rise Rail Station",
+            "Kensal Green Rail Station",
+            "Kent House Rail Station",
+            "Kettering Rail Station",
+            "Kew Bridge Rail Station",
+            "Kew Gardens Rail Station",
+            "Keyham Rail Station",
+            "Keynsham Rail Station",
+            "Kinghorn Rail Station",
+            "Kings Park Rail Station",
+            "Kingston Rail Station",
+            "Kingussie Rail Station",
+            "Kingswood Rail Station",
+            "Kidbrooke Rail Station",
+            "Kidsgrove Rail Station",
+            "Kidwelly Rail Station",
+            "Kildonan Rail Station",
+            "Kilgetty Rail Station",
+            "Kilmarnock Rail Station",
+            "Kilwinning Rail Station",
+            "Kinbrace Rail Station",
+            "Kingham Rail Station",
+            "Kings Sutton Rail Station",
+            "Kirby Cross Rail Station",
+            "Kirkham & Wesham Rail Station",
+            "Kirkby Rail Station",
+            "Kirk Sandall Rail Station",
+            "Kilburn High Road Rail Station",
+            "Kildale Rail Station",
+            "Kings Langley Rail Station",
+            "Kilmaurs Rail Station",
+            "Kilpatrick Rail Station",
+            "Kings Lynn Rail Station",
+            "Kempston Hardwick Rail Station",
+            "Kempton Park Rail Station",
+            "Kemsley Rail Station",
+            "Knockholt Rail Station",
+            "Knebworth Rail Station",
+            "Kingsknowe Rail Station",
+            "London Kings Cross Rail Station",
+            "Knighton Rail Station",
+            "Kenley Rail Station",
+            "Kings Norton Rail Station",
+            "Knaresborough Rail Station",
+            "Kintbury Rail Station",
+            "Knottingley Rail Station",
+            "Kentish Town Rail Station",
+            "Kentish Town West Rail Station",
+            "Knucklas Rail Station",
+            "Knutsford Rail Station",
+            "Kings Nympton Rail Station",
+            "Kirkby-in-Furness Rail Station",
+            "Kirkconnel Rail Station",
+            "Kirkdale Rail Station",
+            "Kirkhill Rail Station",
+            "Kirknewton Rail Station",
+            "Kirkstall Forge Rail Station",
+            "Kirkwood Rail Station",
+            "Kirkby in Ashfield Rail Station",
+            "Kirton Lindsey Rail Station",
+            "Kearsney Rail Station",
+            "Kirkby Stephen Rail Station",
+            "Kents Bank Rail Station",
+            "Kenton Rail Station",
+            "Kiveton Bridge Rail Station",
+            "Kiveton Park Rail Station",
+            "Kyle of Lochalsh Rail Station",
+            "Ladybank Rail Station",
+            "Laindon Rail Station",
+            "Lairg Rail Station",
+            "Lake (Isle of Wight) Rail Station",
+            "Lakenheath Rail Station",
+            "Gilshochill Rail Station",
+            "Lamphey Rail Station",
+            "Lanark Rail Station",
+            "Lancaster Rail Station",
+            "Llandudno Junction Rail Station",
+            "Llanfairfechan Rail Station",
+            "Langley (Berks) Rail Station",
+            "Lapford Rail Station",
+            "Lapworth Rail Station",
+            "Larbert Rail Station",
+            "Largs Rail Station",
+            "Larkhall Rail Station",
+            "Lawrence Hill Rail Station",
+            "Layton (Lancs) Rail Station",
+            "Lazonby & Kirkoswald Rail Station",
+            "Llanbedr Rail Station",
+            "Loughborough Junction Rail Station",
+            "Loch Eil Outward Bound Rail Station",
+            "Lochailort Rail Station",
+            "Lichfield City Rail Station",
+            "Locheilside Rail Station",
+            "Lochgelly Rail Station",
+            "Lochluichart Rail Station",
+            "Lichfield Trent Valley High Level Rail Station",
+            "Lichfield Trent Valley Rail Station",
+            "Lochwinnoch Rail Station",
+            "Lockerbie Rail Station",
+            "Ledbury Rail Station",
+            "Lidlington Rail Station",
+            "Ladywell Rail Station",
+            "Leagrave Rail Station",
+            "Lea Hall Rail Station",
+            "Lealholm Rail Station",
+            "Leasowe Rail Station",
+            "Leeds Rail Station",
+            "Lee Rail Station",
+            "Lea Green Rail Station",
+            "Leigh (Kent) Rail Station",
+            "Lelant Rail Station",
+            "Lelant Saltings Rail Station",
+            "Lenham Rail Station",
+            "Lenzie Rail Station",
+            "Leominster Rail Station",
+            "Leicester Rail Station",
+            "Leatherhead Rail Station",
+            "Leuchars Rail Station",
+            "Lewes Rail Station",
+            "Lewisham Rail Station",
+            "Leyland Rail Station",
+            "Leyton Midland Road Rail Station",
+            "Llanfairpwll Rail Station",
+            "Langbank Rail Station",
+            "Longfield Rail Station",
+            "Langho Rail Station",
+            "Langley Mill Rail Station",
+            "Langley Green Rail Station",
+            "Langwith - Whaley Thorns Rail Station",
+            "Lingwood Rail Station",
+            "Leigh-on-Sea Rail Station",
+            "Littlehaven Rail Station",
+            "Limehouse Rail Station",
+            "Lincoln Central Rail Station",
+            "Lingfield Rail Station",
+            "Liphook Rail Station",
+            "Liskeard Rail Station",
+            "Liss Rail Station",
+            "Littleborough Rail Station",
+            "London Liverpool Street Rail Station",
+            "Lockwood Rail Station",
+            "Llandaf Rail Station",
+            "Llanbradach Rail Station",
+            "Llanbister Road Rail Station",
+            "Llandeilo Rail Station",
+            "Llandrindod Rail Station",
+            "Llandovery Rail Station",
+            "Llandybie Rail Station",
+            "Llangadog Rail Station",
+            "Llangammarch Rail Station",
+            "Llangennech Rail Station",
+            "Llangynllo Rail Station",
+            "Llanharan Rail Station",
+            "Llanhilleth Rail Station",
+            "Llanishen Rail Station",
+            "Llanelli Rail Station",
+            "Llansamlet Rail Station",
+            "Llantwit Major Rail Station",
+            "Llanwrda Rail Station",
+            "Llanwrtyd Rail Station",
+            "Llwynypia Rail Station",
+            "Leamington Spa Rail Station",
+            "Lymington Pier Rail Station",
+            "Lymington Town Rail Station",
+            "Llanaber Rail Station",
+            "Lancing Rail Station",
+            "Llandecwyn Rail Station",
+            "London Bridge Rail Station",
+            "London Bridge Rail Station",
+            "Llandanwg Rail Station",
+            "Landywood Rail Station",
+            "Long Buckby Rail Station",
+            "Longcross Rail Station",
+            "Long Eaton Rail Station",
+            "Longniddry Rail Station",
+            "Longport Rail Station",
+            "Linlithgow Rail Station",
+            "Llanrwst Rail Station",
+            "Langside Rail Station",
+            "Longton Rail Station",
+            "Loch Awe Rail Station",
+            "Loughborough (Leics) Rail Station",
+            "Longbridge Rail Station",
+            "London Fields Rail Station",
+            "Longbeck Rail Station",
+            "Looe Rail Station",
+            "Lostock Rail Station",
+            "Lowdham Rail Station",
+            "Lowestoft Rail Station",
+            "Long Preston Rail Station",
+            "London Road (Guildford) Rail Station",
+            "Laurencekirk Rail Station",
+            "Lostock Gralam Rail Station",
+            "Lostock Hall Rail Station",
+            "Lostwithiel Rail Station",
+            "Lower Sydenham Rail Station",
+            "Letchworth Rail Station",
+            "Lisvane & Thornhill Rail Station",
+            "Littlehampton Rail Station",
+            "Little Kimble Rail Station",
+            "Littleport Rail Station",
+            "Little Sutton Rail Station",
+            "Leighton Buzzard Rail Station",
+            "Ludlow Rail Station",
+            "Llandudno Rail Station",
+            "Luton Airport Parkway Rail Station",
+            "Luton Rail Station",
+            "Luxulyan Rail Station",
+            "Levenshulme Rail Station",
+            "Livingston North Rail Station",
+            "Livingston South Rail Station",
+            "Liverpool Central Rail Station",
+            "Liverpool Central Loop Line",
+            "Liverpool Lime Street Rail Station",
+            "Liverpool Lime Street Low Level Rail Station",
+            "Liverpool South Parkway Rail Station",
+            "Langwathby Rail Station",
+            "Lydney Rail Station",
+            "Lye (West Midlands) Rail Station",
+            "Lympstone Commando Rail Station",
+            "Lympstone Village Rail Station",
+            "Ashurst New Forest Rail Station",
+            "Llwyngwril Rail Station",
+            "Lytham Rail Station",
+            "Leytonstone High Road Rail Station",
+            "Macclesfield Rail Station",
+            "Maesteg (Ewenny Road) Rail Station",
+            "Maesteg Rail Station",
+            "Maghull Rail Station",
+            "Malden Manor Rail Station",
+            "Malton Rail Station",
+            "Manea Rail Station",
+            "Manningtree Rail Station",
+            "Manors Rail Station",
+            "Manor Park Rail Station",
+            "Marden Rail Station",
+            "Margate Rail Station",
+            "Marlow Rail Station",
+            "Marple Rail Station",
+            "Marton Rail Station",
+            "London Marylebone Rail Station",
+            "Mountain Ash Rail Station",
+            "Matlock Bath Rail Station",
+            "Matlock Rail Station",
+            "Maxwell Park Rail Station",
+            "Maybole Rail Station",
+            "Maze Hill Rail Station",
+            "Millbrook (Hants) Rail Station",
+            "Micheldever Rail Station",
+            "Machynlleth Rail Station",
+            "Micklefield Rail Station",
+            "Middlesbrough Rail Station",
+            "Maidenhead Rail Station",
+            "Maiden Newton Rail Station",
+            "Middlewood Rail Station",
+            "Meadowhall Rail Station",
+            "Meldreth Rail Station",
+            "Melksham Rail Station",
+            "Meols Rail Station",
+            "Melton (Suffolk) Rail Station",
+            "Menheniot Rail Station",
+            "Menston Rail Station",
+            "Meols Cop Rail Station",
+            "Meopham Rail Station",
+            "Merstham Rail Station",
+            "Merthyr Vale Rail Station",
+            "Merthyr Tydfil Rail Station",
+            "Mexborough Rail Station",
+            "Morfa Mawddach Rail Station",
+            "Mansfield Rail Station",
+            "Mansfield Woodhouse Rail Station",
+            "Martins Heron Rail Station",
+            "Midgham Rail Station",
+            "Milliken Park Rail Station",
+            "Minffordd Ffestiniog Railway Station",
+            "Minffordd Rail Station",
+            "Moreton-in-Marsh Rail Station",
+            "Minster Rail Station",
+            "Mirfield Rail Station",
+            "Mistley Rail Station",
+            "Mitcham Junction Rail Station",
+            "Markinch Rail Station",
+            "Milton Keynes Central Rail Station",
+            "Market Rasen Rail Station",
+            "Mallaig Rail Station",
+            "Millbrook (Beds) Rail Station",
+            "Mouldsworth Rail Station",
+            "Mauldeth Road Rail Station",
+            "Milford Haven Rail Station",
+            "Milford (Surrey) Rail Station",
+            "Milngavie Rail Station",
+            "Mill Hill Broadway Rail Station",
+            "Mill Hill (Lancs) Rail Station",
+            "Millom Rail Station",
+            "Moulsecoomb Rail Station",
+            "Mills Hill (Manchester) Rail Station",
+            "Melton Mowbray Rail Station",
+            "Malvern Link Rail Station",
+            "Martin Mill Rail Station",
+            "Deansgate Rail Station",
+            "Manchester Airport Rail Station",
+            "Manchester Oxford Road Rail Station",
+            "Manchester Piccadilly Rail Station",
+            "Manchester United FC Rail Station",
+            "Manchester Victoria Rail Station",
+            "Monks Risborough Rail Station",
+            "Manorbier Rail Station",
+            "Manor Road Rail Station",
+            "Mount Vernon Rail Station",
+            "Mobberley Rail Station",
+            "Morchard Road Rail Station",
+            "Monifieth Rail Station",
+            "Montpelier Rail Station",
+            "Montrose Rail Station",
+            "Morar Rail Station",
+            "Morecambe Rail Station",
+            "Morden South Rail Station",
+            "Moreton (Dorset) Rail Station",
+            "Moorfields Rail Station",
+            "Moorfields Rail Station",
+            "Moorthorpe Rail Station",
+            "Mossley (Manchester) Rail Station",
+            "Mottisfont & Dunbridge Rail Station",
+            "Motherwell Rail Station",
+            "Mottingham Rail Station",
+            "Moreton (Merseyside) Rail Station",
+            "Motspur Park Rail Station",
+            "Maryport Rail Station",
+            "March Rail Station",
+            "Moorgate Rail Station",
+            "Marks Tey Rail Station",
+            "Market Harborough Rail Station",
+            "Morley Rail Station",
+            "Morpeth Rail Station",
+            "Merryton Rail Station",
+            "Moorside Rail Station",
+            "Marske Rail Station",
+            "Marsden Rail Station",
+            "Marston Green Rail Station",
+            "Mortlake Rail Station",
+            "Maryhill Rail Station",
+            "Maryland Rail Station",
+            "Musselburgh Rail Station",
+            "Moses Gate Rail Station",
+            "Mossley Hill Rail Station",
+            "Moss Side Rail Station",
+            "Mosspark Rail Station",
+            "Moston Rail Station",
+            "Maidstone Barracks Rail Station",
+            "Maidstone East Rail Station",
+            "Maidstone West Rail Station",
+            "Mount Florida Rail Station",
+            "Metheringham Rail Station",
+            "Mortimer Rail Station",
+            "Muirend Rail Station",
+            "Muir of Ord Rail Station",
+            "Mytholmroyd Rail Station",
+            "Newton Abbot Rail Station",
+            "Nafferton Rail Station",
+            "Nailsea & Backwell Rail Station",
+            "Nairn Rail Station",
+            "Nantwich Rail Station",
+            "Narborough Rail Station",
+            "Navigation Road Rail Station",
+            "New Barnet Rail Station",
+            "New Beckenham Rail Station",
+            "North Berwick Rail Station",
+            "Newbridge Rail Station",
+            "New Brighton Rail Station",
+            "Newcraighall Rail Station",
+            "North Dulwich Rail Station",
+            "Neath Rail Station",
+            "Needham Market Rail Station",
+            "Neilston Rail Station",
+            "Neston Rail Station",
+            "Nethertown Rail Station",
+            "Netley Rail Station",
+            "Newark North Gate Rail Station",
+            "Newbury Racecourse Rail Station",
+            "Newbury Rail Station",
+            "New Clee Rail Station",
+            "Newhaven Harbour Rail Station",
+            "Newhaven Town Rail Station",
+            "New Lane Rail Station",
+            "New Malden Rail Station",
+            "Newington Rail Station",
+            "Newport (Essex) Rail Station",
+            "Newquay Rail Station",
+            "New Southgate Rail Station",
+            "New Cross Gate ELL Rail Station",
+            "New Cross Gate Rail Station",
+            "New Holland Rail Station",
+            "Ninian Park Rail Station",
+            "Nitshill Rail Station",
+            "North Llanrwst Rail Station",
+            "Northallerton Rail Station",
+            "Nelson Rail Station",
+            "Northumberland Park Rail Station",
+            "New Milton Rail Station",
+            "Northampton Rail Station",
+            "Nuneaton Rail Station",
+            "Norbury Rail Station",
+            "Normanton Rail Station",
+            "Norwood Junction Rail Station",
+            "New Pudsey Rail Station",
+            "North Queensferry Rail Station",
+            "Narberth Rail Station",
+            "Norbiton Rail Station",
+            "Norwich Rail Station",
+            "Normans Bay Rail Station",
+            "Northfield Rail Station",
+            "Northfleet Rail Station",
+            "North Sheen Rail Station",
+            "Newton St Cyres Rail Station",
+            "North Camp Rail Station",
+            "Northolt Park Rail Station",
+            "Netherfield Rail Station",
+            "Northwich Rail Station",
+            "Norton Bridge Rail Station",
+            "Nottingham Rail Station",
+            "North Road Rail Station",
+            "Nunhead Rail Station",
+            "Nunthorpe Rail Station",
+            "Nutbourne Rail Station",
+            "Nutfield Rail Station",
+            "North Walsham Rail Station",
+            "New Cumnock Rail Station",
+            "Newcourt Rail Station",
+            "New Cross ELL Rail Station",
+            "New Cross Rail Station",
+            "Newcastle Rail Station",
+            "New Eltham Rail Station",
+            "North Wembley Rail Station",
+            "New Hythe Rail Station",
+            "New Mills Central Rail Station",
+            "New Mills Newtown Rail Station",
+            "Newmarket Rail Station",
+            "Newport (S Wales) Rail Station",
+            "Newark Castle Rail Station",
+            "Newstead Rail Station",
+            "Newton Aycliffe Rail Station",
+            "Newtongrange Rail Station",
+            "Newton for Hyde Rail Station",
+            "Newton (S Lanarks) Rail Station",
+            "Newtonmore Rail Station",
+            "Newton-le-Willows Rail Station",
+            "Newton-on-Ayr Rail Station",
+            "Newtown (Powys) Rail Station",
+            "Oakham Rail Station",
+            "Oakengates Rail Station",
+            "Oban Rail Station",
+            "Ockendon Rail Station",
+            "Ockley Rail Station",
+            "Okehampton Rail Station",
+            "Oakleigh Park Rail Station",
+            "Oldfield Park Rail Station",
+            "Old Roan Rail Station",
+            "Old Street Rail Station",
+            "Old Hill Rail Station",
+            "Olton Rail Station",
+            "Overpool Rail Station",
+            "Ore Rail Station",
+            "Orrell Park Rail Station",
+            "Ormskirk Rail Station",
+            "Orpington Rail Station",
+            "Orrell Rail Station",
+            "Otford Rail Station",
+            "Oulton Broad North Rail Station",
+            "Oulton Broad South Rail Station",
+            "Outwood Rail Station",
+            "Overton Rail Station",
+            "Oxenholme Lake District Rail Station",
+            "Oxford Rail Station",
+            "Oxford Parkway Rail Station",
+            "Oxshott Rail Station",
+            "Oxted Rail Station",
+            "Oxted Rail Station",
+            "Padgate Rail Station",
+            "London Paddington Rail Station",
+            "Paignton Rail Station",
+            "Palmers Green Rail Station",
+            "Pangbourne Rail Station",
+            "Pannal Rail Station",
+            "Parbold Rail Station",
+            "Par Rail Station",
+            "Parson Street Rail Station",
+            "Parton Rail Station",
+            "Patchway Rail Station",
+            "Patterton Rail Station",
+            "Peterborough Rail Station",
+            "Pemberton Rail Station",
+            "Portchester Rail Station",
+            "Queens Road Peckham Rail Station",
+            "Peckham Rye Rail Station",
+            "Peartree Rail Station",
+            "Pegswood Rail Station",
+            "Pembroke Dock Rail Station",
+            "Pembroke Rail Station",
+            "Penarth Rail Station",
+            "Pencoed Rail Station",
+            "Penge West Rail Station",
+            "Pengam Rail Station",
+            "Penhelig Rail Station",
+            "Penistone Rail Station",
+            "Penmere Rail Station",
+            "Penychain Rail Station",
+            "Penzance Rail Station",
+            "Perranwell Rail Station",
+            "Pershore Rail Station",
+            "Perth Rail Station",
+            "Petts Wood Rail Station",
+            "Pewsey Rail Station",
+            "Portsmouth Harbour Rail Station",
+            "Priesthill & Darnley Rail Station",
+            "Pilning Rail Station",
+            "Pinhoe Rail Station",
+            "Pitsea Rail Station",
+            "Peckham Rye Rail Station",
+            "Park Street Rail Station",
+            "Paddock Wood Rail Station",
+            "Pollokshields East Rail Station",
+            "Pollokshields West Rail Station",
+            "Pollokshaws East Rail Station",
+            "Pollokshaws West Rail Station",
+            "Plumpton Rail Station",
+            "Plumstead Rail Station",
+            "Plockton Rail Station",
+            "Pleasington Rail Station",
+            "Portlethen Rail Station",
+            "Poulton-le-Fylde Rail Station",
+            "Pluckley Rail Station",
+            "Plumley Rail Station",
+            "Plymouth Rail Station",
+            "Pembrey & Burry Port Rail Station",
+            "Penally Rail Station",
+            "Ponders End Rail Station",
+            "Penge East Rail Station",
+            "Penkridge Rail Station",
+            "Penmaenmawr Rail Station",
+            "Penrhiwceiber Rail Station",
+            "Penshurst Rail Station",
+            "Pontarddulais Rail Station",
+            "Penrith North Lakes Rail Station",
+            "Pontlottyn Rail Station",
+            "Pentre-Bach Rail Station",
+            "Pantyffynnon Rail Station",
+            "Penyffordd Rail Station",
+            "Pont-y-Pant Rail Station",
+            "Pokesdown Rail Station",
+            "Polegate Rail Station",
+            "Polmont Rail Station",
+            "Polsloe Bridge Rail Station",
+            "Pontypool & New Inn Rail Station",
+            "Pontyclun Rail Station",
+            "Poole Rail Station",
+            "Poppleton Rail Station",
+            "Porth Rail Station",
+            "Possilpark & Parkhouse Rail Station",
+            "Potters Bar Rail Station",
+            "Poynton Rail Station",
+            "Prestonpans Rail Station",
+            "Prees Rail Station",
+            "Prescot Rail Station",
+            "Penrhyndeudraeth Rail Station",
+            "Princes Risborough Rail Station",
+            "Prittlewell Rail Station",
+            "Harwich International Rail Station",
+            "Preston Park Rail Station",
+            "Preston Rail Station",
+            "Porthmadog Harbour Ffestiniog Railway Station",
+            "Porthmadog Rail Station",
+            "Portsmouth Arms Rail Station",
+            "Prudhoe Rail Station",
+            "Perry Barr Rail Station",
+            "Penryn Rail Station",
+            "Prestbury Rail Station",
+            "Portsmouth & Southsea Rail Station",
+            "Portslade Rail Station",
+            "Port Sunlight Rail Station",
+            "Paisley Canal Rail Station",
+            "Paisley Gilmour Street Rail Station",
+            "Paisley St James Rail Station",
+            "Pensarn (Gwynedd) Rail Station",
+            "Prestatyn Rail Station",
+            "Parkstone (Dorset) Rail Station",
+            "Port Talbot Parkway Rail Station",
+            "Partick Rail Station",
+            "Pontefract Baghill Rail Station",
+            "Pontefract Monkhill Rail Station",
+            "Pontefract Tanshelf Rail Station",
+            "Port Glasgow Rail Station",
+            "Pitlochry Rail Station",
+            "Patricroft Rail Station",
+            "Petersfield Rail Station",
+            "Pontypridd Rail Station",
+            "Pulborough Rail Station",
+            "Purfleet Rail Station",
+            "Purley Rail Station",
+            "Purley Oaks Rail Station",
+            "Putney Rail Station",
+            "Pevensey & Westham Rail Station",
+            "Pevensey Bay Rail Station",
+            "Prestwick Rail Station",
+            "Prestwick Intl Airport Rail Station",
+            "Pwllheli Rail Station",
+            "Polesworth Rail Station",
+            "Pen-y-Bont Rail Station",
+            "Pyle Rail Station",
+            "Quintrell Downs Rail Station",
+            "Queens Park (London) Rail Station",
+            "Queenstown Road (Battersea) Rail Station",
+            "Quakers Yard Rail Station",
+            "Queenborough Rail Station",
+            "Queens Park (Glasgow) Rail Station",
+            "Radlett Rail Station",
+            "Radcliffe (Notts) Rail Station",
+            "Radyr Rail Station",
+            "Rainham (Kent) Rail Station",
+            "Ramsgate Rail Station",
+            "Rannoch Rail Station",
+            "Rauceby Rail Station",
+            "Rawcliffe Rail Station",
+            "Rayleigh Rail Station",
+            "Raynes Park Rail Station",
+            "Ravensbourne Rail Station",
+            "Robertsbridge Rail Station",
+            "Rochdale Rail Station",
+            "Roche Rail Station",
+            "Rochester Rail Station",
+            "Rickmansworth Rail Station",
+            "Rectory Road Rail Station",
+            "Reedham (Norfolk) Rail Station",
+            "Redditch Rail Station",
+            "Radley Rail Station",
+            "Riddlesdown Rail Station",
+            "Reading Rail Station",
+            "Reading Rail Station",
+            "Reading Rail Station",
+            "Redbridge (Hants) Rail Station",
+            "Redcar Central Rail Station",
+            "Redcar East Rail Station",
+            "British Steel Redcar Rail Station",
+            "Reading West Rail Station",
+            "Redhill Rail Station",
+            "Reddish North Rail Station",
+            "Reddish South Rail Station",
+            "Redland Rail Station",
+            "Redruth Rail Station",
+            "Reedham (Surrey) Rail Station",
+            "Reigate Rail Station",
+            "Renton Rail Station",
+            "Roughton Road Rail Station",
+            "Ravenglass Rail Station",
+            "Rainhill Rail Station",
+            "Rhiwbina Rail Station",
+            "Rhoose Rail Station",
+            "Rhosneigr Rail Station",
+            "Rhyl Rail Station",
+            "Rhymney Rail Station",
+            "Ribblehead Rail Station",
+            "Rice Lane Rail Station",
+            "Richmond (London) Rail Station",
+            "Richmond NLL Rail Station",
+            "Ridgmont Rail Station",
+            "Riding Mill Rail Station",
+            "Risca & Pontymister Rail Station",
+            "Rishton Rail Station",
+            "Ramsgreave & Wilpshire Rail Station",
+            "Rainford Rail Station",
+            "Rainham (London) Rail Station",
+            "Roby Rail Station",
+            "Rochford Rail Station",
+            "Rock Ferry Rail Station",
+            "Rogart Rail Station",
+            "Rogerstone Rail Station",
+            "Rose Hill Marple Rail Station",
+            "Rolleston Rail Station",
+            "Roman Bridge Rail Station",
+            "Romford Rail Station",
+            "Romiley Rail Station",
+            "Romsey Rail Station",
+            "Roose Rail Station",
+            "Rose Grove Rail Station",
+            "Rosyth Rail Station",
+            "Rotherham Central Rail Station",
+            "Rowley Regis Rail Station",
+            "Roy Bridge Rail Station",
+            "Roydon Rail Station",
+            "Royston Rail Station",
+            "Ruskington Rail Station",
+            "Retford Rail Station",
+            "Retford Low Level Rail Station",
+            "Rotherhithe Rail Station",
+            "Rutherglen Rail Station",
+            "Ravensthorpe Rail Station",
+            "Ruabon Rail Station",
+            "Rufford Rail Station",
+            "Rugby Rail Station",
+            "Rugeley Trent Valley Rail Station",
+            "Rugeley Town Rail Station",
+            "Runcorn Rail Station",
+            "Runcorn East Rail Station",
+            "Ruswarp Rail Station",
+            "Rowlands Castle Rail Station",
+            "Ryde Esplanade Rail Station",
+            "Ryde Pier Head Rail Station",
+            "Ryder Brow Rail Station",
+            "Ryde St Johns Road Rail Station",
+            "Rye Rail Station",
+            "Rye House Rail Station",
+            "Sandal & Agbrigg Rail Station",
+            "South Acton Rail Station",
+            "St Andrews Road Rail Station",
+            "Saltaire Rail Station",
+            "Salfords (Surrey) Rail Station",
+            "Salhouse Rail Station",
+            "Salwick Rail Station",
+            "Sampford Courtenay Rail Station",
+            "Sandhills Rail Station",
+            "Sankey for Penketh Rail Station",
+            "Sanquhar Rail Station",
+            "Sarn Rail Station",
+            "Saltash Rail Station",
+            "Saxilby Rail Station",
+            "Sandbach Rail Station",
+            "Sawbridgeworth Rail Station",
+            "Southbourne Rail Station",
+            "South Bermondsey Rail Station",
+            "Saltburn Rail Station",
+            "Stallingborough Rail Station",
+            "Southbury Rail Station",
+            "Stalybridge Rail Station",
+            "Scarborough Rail Station",
+            "Scunthorpe Rail Station",
+            "South Croydon Rail Station",
+            "Scotscalder Rail Station",
+            "Scotstounhill Rail Station",
+            "Silverdale Rail Station",
+            "Sudbury Hill Harrow Rail Station",
+            "Sudbury & Harrow Road Rail Station",
+            "Sandling Rail Station",
+            "Swindon Rail Station",
+            "Sandplace Rail Station",
+            "Saundersfoot Rail Station",
+            "Sanderstead Rail Station",
+            "Seaburn Rail Station",
+            "Seaford Rail Station",
+            "Seaham Rail Station",
+            "Seamer Rail Station",
+            "Seascale Rail Station",
+            "Seer Green Rail Station",
+            "Sellafield Rail Station",
+            "Selby Rail Station",
+            "Selhurst Rail Station",
+            "Selling Rail Station",
+            "South Elmsall Rail Station",
+            "Selly Oak Rail Station",
+            "Sea Mills Rail Station",
+            "Southease Rail Station",
+            "Seaton Carew Rail Station",
+            "Settle Rail Station",
+            "Seven Sisters Rail Station",
+            "Severn Tunnel Junction Rail Station",
+            "Seaforth & Litherland Rail Station",
+            "South Greenford Rail Station",
+            "Sugar Loaf Rail Station",
+            "Shadwell Rail Station",
+            "Shalford (Surrey) Rail Station",
+            "Shirehampton Rail Station",
+            "Shanklin Rail Station",
+            "Shawford Rail Station",
+            "Sherburn-in-Elmet Rail Station",
+            "Shoeburyness Rail Station",
+            "Shildon Rail Station",
+            "Sheffield Rail Station",
+            "Shelford (Cambs) Rail Station",
+            "Shenfield Rail Station",
+            "Shepperton Rail Station",
+            "Sherborne Rail Station",
+            "Shifnal Rail Station",
+            "Shieldmuir Rail Station",
+            "Shirley Rail Station",
+            "St Helier (London) Rail Station",
+            "South Hampstead Rail Station",
+            "Shenstone Rail Station",
+            "Sholing Rail Station",
+            "Shotton High Level Rail Station",
+            "Shotton Rail Station",
+            "Shepherds Bush Rail Station",
+            "Stanhope Rail Station",
+            "Shiplake Rail Station",
+            "Shepreth Rail Station",
+            "Shipley Rail Station",
+            "Shirebrook Rail Station",
+            "Shoreditch High Street Rail Station",
+            "Sheringham Rail Station",
+            "Shoreham-by-Sea (Sussex) Rail Station",
+            "Shireoaks Rail Station",
+            "Shoreham (Kent) Rail Station",
+            "Sheerness-on-Sea Rail Station",
+            "Shortlands Rail Station",
+            "Shrewsbury Rail Station",
+            "Shettleston Rail Station",
+            "Shotts Rail Station",
+            "Stonehaven Rail Station",
+            "Shawfair Rail Station",
+            "Shawlands Rail Station",
+            "Sidcup Rail Station",
+            "Silecroft Rail Station",
+            "Sileby Rail Station",
+            "Singer Rail Station",
+            "Silver Street Rail Station",
+            "Skegness Rail Station",
+            "South Kenton Rail Station",
+            "Skewen Rail Station",
+            "Skipton Rail Station",
+            "Silkstone Common Rail Station",
+            "Slade Green Rail Station",
+            "Slateford Rail Station",
+            "Saltcoats Rail Station",
+            "Sleaford Rail Station",
+            "Sleights Rail Station",
+            "Stanford-le-Hope Rail Station",
+            "Salford Crescent Rail Station",
+            "Salford Central Rail Station",
+            "Slough Rail Station",
+            "Salisbury Rail Station",
+            "Slaithwaite Rail Station",
+            "Saltmarshe Rail Station",
+            "Smallbrook Junction Rail Station",
+            "St Margarets (Herts) Rail Station",
+            "Smithy Bridge Rail Station",
+            "South Merton Rail Station",
+            "South Milford Rail Station",
+            "Southminster Rail Station",
+            "Small Heath Rail Station",
+            "Somerleyton Rail Station",
+            "Smethwick Rolfe Street Rail Station",
+            "Snaith Rail Station",
+            "Sandhurst Rail Station",
+            "Sandown Rail Station",
+            "Sundridge Park Rail Station",
+            "Sunderland Rail Station",
+            "St Peters Rail Station",
+            "Sandwell & Dudley Rail Station",
+            "Sandy Rail Station",
+            "Stonegate Rail Station",
+            "Snodland Rail Station",
+            "Snowdown Rail Station",
+            "Stratford-upon-Avon Rail Station",
+            "Stratford-upon-Avon Parkway Rail Station",
+            "Sole Street Rail Station",
+            "Solihull Rail Station",
+            "Southampton Town Quay",
+            "Southampton Central Rail Station",
+            "Southampton Airport Parkway Rail Station",
+            "Southport Rail Station",
+            "Spalding Rail Station",
+            "Spondon Rail Station",
+            "Spean Bridge Rail Station",
+            "Springfield Rail Station",
+            "Shippea Hill Rail Station",
+            "Staplehurst Rail Station",
+            "Spital Rail Station",
+            "Shepley Rail Station",
+            "Spooner Row Rail Station",
+            "Spring Road Rail Station",
+            "Springburn Rail Station",
+            "Shipton Rail Station",
+            "Squires Gate Rail Station",
+            "South Ruislip Rail Station",
+            "Southsea Hoverport",
+            "Stafford Rail Station",
+            "Staines Rail Station",
+            "St Albans City Rail Station",
+            "St Albans Abbey Rail Station",
+            "Stamford Rail Station",
+            "Stansted Airport Rail Station",
+            "Stansted Mountfitchet Rail Station",
+            "Starbeck Rail Station",
+            "St Annes-on-the-Sea Rail Station",
+            "St Austell Rail Station",
+            "Staveley Rail Station",
+            "St Budeaux Ferry Road Rail Station",
+            "St Budeaux Victoria Road Rail Station",
+            "St Bees Rail Station",
+            "Stocksfield Rail Station",
+            "St Columb Road Rail Station",
+            "Strathcarron Rail Station",
+            "Stone Crossing Rail Station",
+            "St Denys Rail Station",
+            "Stechford Rail Station",
+            "Steeton & Silsden Rail Station",
+            "Stepps Rail Station",
+            "St Erth Rail Station",
+            "Stevenage Rail Station",
+            "Stewarton Rail Station",
+            "Stratford (London) Rail Station",
+            "Stratford International Rail Station",
+            "St Germans Rail Station",
+            "Southall Rail Station",
+            "South Bank Rail Station",
+            "Southend Central Rail Station",
+            "Southend Airport Rail Station",
+            "South Gyle Rail Station",
+            "St Helens Central Rail Station",
+            "St Helens Junction Rail Station",
+            "Southend East Rail Station",
+            "Streethouse Rail Station",
+            "Southend Victoria Rail Station",
+            "Southwick Rail Station",
+            "Stirling Rail Station",
+            "St Ives (Cornwall) Rail Station",
+            "St James Park (Devon) Rail Station",
+            "St James Street (London) Rail Station",
+            "St Johns (London) Rail Station",
+            "St Keyne Wishing Well Halt (Rail Station)",
+            "Stoke Mandeville Rail Station",
+            "Stoke Newington Rail Station",
+            "Stockport Rail Station",
+            "Stocksmoor Rail Station",
+            "Stoneleigh Rail Station",
+            "St Leonards Warrior Square Rail Station",
+            "St Michaels Rail Station",
+            "Stamford Hill Rail Station",
+            "St Margarets (London) Rail Station",
+            "St Mary Cray Rail Station",
+            "Stonebridge Park Rail Station",
+            "St Neots Rail Station",
+            "Sittingbourne Rail Station",
+            "Stonehouse Rail Station",
+            "Stanlow & Thornton Rail Station",
+            "Stockton Rail Station",
+            "Stoke-on-Trent Rail Station",
+            "Stone Rail Station",
+            "South Tottenham Rail Station",
+            "Stow Rail Station",
+            "London St Pancras International Rail Station",
+            "Stapleton Road Rail Station",
+            "London St Pancras International Rail Station",
+            "London St Pancras International LL Rail Station",
+            "Stourbridge Junction Rail Station",
+            "Stourbridge Junction Rail Station",
+            "Stourbridge Town Rail Station",
+            "Starcross Rail Station",
+            "Streatham Rail Station",
+            "Streatham Common Rail Station",
+            "Streatham Hill Rail Station",
+            "Strines Rail Station",
+            "Stromeferry Rail Station",
+            "Stranraer Rail Station",
+            "Strood Rail Station",
+            "Stroud (Glos) Rail Station",
+            "Strawberry Hill Rail Station",
+            "Sutton Parkway Rail Station",
+            "Sturry Rail Station",
+            "Stevenston Rail Station",
+            "Stowmarket Rail Station",
+            "Stewartby Rail Station",
+            "Styal Rail Station",
+            "Sudbury (Suffolk) Rail Station",
+            "Summerston Rail Station",
+            "Sunbury Rail Station",
+            "Saunderton Rail Station",
+            "Sunningdale Rail Station",
+            "Sunnymeads Rail Station",
+            "Surbiton Rail Station",
+            "Surrey Quays Rail Station",
+            "Sutton Coldfield Rail Station",
+            "Sutton (London) Rail Station",
+            "Sutton Common Rail Station",
+            "Seven Kings Rail Station",
+            "Sevenoaks Rail Station",
+            "Severn Beach Rail Station",
+            "Swale Rail Station",
+            "Swansea Rail Station",
+            "Sway Rail Station",
+            "Sandwich Rail Station",
+            "Shepherds Well Rail Station",
+            "Swinton (South Yorks) Rail Station",
+            "Swanley Rail Station",
+            "Swinderby Rail Station",
+            "Swanscombe Rail Station",
+            "Swinton (Manchester) Rail Station",
+            "Swanwick Rail Station",
+            "Sowerby Bridge Rail Station",
+            "Swineshead Rail Station",
+            "Swaythling Rail Station",
+            "Saxmundham Rail Station",
+            "Sydenham Rail Station",
+            "Sydenham Hill Rail Station",
+            "Syon Lane Rail Station",
+            "Syston Rail Station",
+            "Thornaby Rail Station",
+            "Tackley Rail Station",
+            "Tadworth Rail Station",
+            "Taffs Well Rail Station",
+            "Tain Rail Station",
+            "Talybont Rail Station",
+            "Tal-y-Cafn Rail Station",
+            "Tame Bridge Parkway Rail Station",
+            "Tan-y-Bwlch Ffestiniog Railway Station",
+            "Taplow Rail Station",
+            "Tattenham Corner Rail Station",
+            "Taunton Rail Station",
+            "Taynuilt Rail Station",
+            "Thorpe Culvert Rail Station",
+            "Thames Ditton Rail Station",
+            "Teddington Rail Station",
+            "Teignmouth Rail Station",
+            "Telford Central Rail Station",
+            "Tenby Rail Station",
+            "Teynham Rail Station",
+            "Thornton Abbey Rail Station",
+            "Thorntonhall Rail Station",
+            "Thatcham Rail Station",
+            "Three Bridges Rail Station",
+            "Theobalds Grove Rail Station",
+            "Theale Rail Station",
+            "The Lakes Rail Station",
+            "Thetford Rail Station",
+            "Thornford Rail Station",
+            "Thatto Heath Rail Station",
+            "Thirsk Rail Station",
+            "Thornliebank Rail Station",
+            "Three Oaks Rail Station",
+            "Thorpe Bay Rail Station",
+            "Thorne North Rail Station",
+            "Thorne South Rail Station",
+            "Thorpe-le-Soken Rail Station",
+            "Thurgarton Rail Station",
+            "Thornton Heath Rail Station",
+            "Thurso Rail Station",
+            "Thurston Rail Station",
+            "Tile Hill Rail Station",
+            "Tilehurst Rail Station",
+            "Tipton Rail Station",
+            "Tir-phil Rail Station",
+            "Tisbury Rail Station",
+            "Tiverton Parkway Rail Station",
+            "Tilbury Town Rail Station",
+            "Talsarnau Rail Station",
+            "Templecombe Rail Station",
+            "Tamworth High Level Rail Station",
+            "Tamworth Rail Station",
+            "Tonypandy Rail Station",
+            "Todmorden Rail Station",
+            "Tolworth Rail Station",
+            "Tonbridge Rail Station",
+            "Tondu Rail Station",
+            "Tonfanau Rail Station",
+            "Ton Pentre Rail Station",
+            "Tooting Rail Station",
+            "Topsham Rail Station",
+            "Torquay Rail Station",
+            "Torre Rail Station",
+            "Totnes Rail Station",
+            "Totton Rail Station",
+            "Town Green Rail Station",
+            "Troed-y-Rhiw Rail Station",
+            "Trefforest Rail Station",
+            "Trehafod Rail Station",
+            "Treherbert Rail Station",
+            "Treorchy Rail Station",
+            "Trafford Park Rail Station",
+            "Trefforest Estate Rail Station",
+            "Trimley Rail Station",
+            "Tring Rail Station",
+            "Thurnscoe Rail Station",
+            "Troon Rail Station",
+            "Truro Rail Station",
+            "Trowbridge Rail Station",
+            "Tees-side Airport Rail Station",
+            "Tottenham Hale Rail Station",
+            "Tulloch Rail Station",
+            "Tulse Hill Rail Station",
+            "Tunbridge Wells Rail Station",
+            "Turkey Street Rail Station",
+            "Tutbury & Hatton Rail Station",
+            "Twickenham Rail Station",
+            "Tweedbank Rail Station",
+            "Twyford Rail Station",
+            "Ty Croes Rail Station",
+            "Ty Glas Rail Station",
+            "Tygwyn Rail Station",
+            "Tyndrum Lower Rail Station",
+            "Upper Tyndrum Rail Station",
+            "Tyseley Rail Station",
+            "Tywyn Rail Station",
+            "Uckfield Rail Station",
+            "Uddingston Rail Station",
+            "Upper Halliford Rail Station",
+            "Ulceby Rail Station",
+            "Ulleskelf Rail Station",
+            "Ulverston Rail Station",
+            "Umberleigh Rail Station",
+            "University Rail Station",
+            "Uphall Rail Station",
+            "Upholland Rail Station",
+            "Upminster Rail Station",
+            "Upminster Rail Station",
+            "Upper Holloway Rail Station",
+            "Upton Rail Station",
+            "Upwey Rail Station",
+            "Urmston Rail Station",
+            "Uttoxeter Rail Station",
+            "Upper Warlingham Rail Station",
+            "Valley Rail Station",
+            "Vauxhall Rail Station",
+            "Vauxhall Rail Station",
+            "London Victoria Rail Station",
+            "London Victoria Rail Station",
+            "Virginia Water Rail Station",
+            "Waddon Rail Station",
+            "Wadhurst Rail Station",
+            "Wainfleet Rail Station",
+            "Wallasey Grove Road Rail Station",
+            "Wallasey Village Rail Station",
+            "West Allerton Rail Station",
+            "Waltham Cross Rail Station",
+            "Walkden Rail Station",
+            "Wallyford Rail Station",
+            "Walmer Rail Station",
+            "Wallington Rail Station",
+            "Walsall Rail Station",
+            "Walton (Merseyside) Rail Station",
+            "Wanborough Rail Station",
+            "Wandsworth Common Rail Station",
+            "Wapping Rail Station",
+            "Ware Rail Station",
+            "Wargrave Rail Station",
+            "Wareham Rail Station",
+            "Warnham Rail Station",
+            "Warwick Rail Station",
+            "Warwick Parkway Rail Station",
+            "Waterbeach Rail Station",
+            "Watford High Street Rail Station",
+            "Watford Junction Rail Station",
+            "Watford North Rail Station",
+            "Watford Junction Rail Station",
+            "Watlington Rail Station",
+            "Watton-at-Stone Rail Station",
+            "London Waterloo Rail Station",
+            "Whatstandwell Rail Station",
+            "Waun-gron Park Rail Station",
+            "West Brompton Rail Station",
+            "West Byfleet Rail Station",
+            "West Calder Rail Station",
+            "Whitechapel Rail Station",
+            "Whitchurch (Shrops) Rail Station",
+            "Wickham Market Rail Station",
+            "Westcliff-on-Sea Rail Station",
+            "Westcombe Park Rail Station",
+            "West Croydon Rail Station",
+            "Walsden Rail Station",
+            "Woodgrange Park Rail Station",
+            "South Woodham Ferrers Rail Station",
+            "Woodhouse Rail Station",
+            "Woodlesford Rail Station",
+            "Woodmansterne Rail Station",
+            "Wood End Rail Station",
+            "Widney Manor Rail Station",
+            "Wimbledon Rail Station",
+            "Widdrington Rail Station",
+            "West Drayton Rail Station",
+            "Wood Street Rail Station",
+            "West Dulwich Rail Station",
+            "Wandsworth Town Rail Station",
+            "West Ealing Rail Station",
+            "Wedgwood Rail Station",
+            "Weeley Rail Station",
+            "Weeton Rail Station",
+            "Welham Green Rail Station",
+            "Welling Rail Station",
+            "Wellington (Shropshire) Rail Station",
+            "Welshpool Rail Station",
+            "Wembley Stadium Rail Station",
+            "Wem Rail Station",
+            "Wemyss Bay Rail Station",
+            "Wendover Rail Station",
+            "Westenhanger Rail Station",
+            "Wennington Rail Station",
+            "Westgate-on-Sea Rail Station",
+            "Westerton Rail Station",
+            "Wetheral Rail Station",
+            "Weybridge Rail Station",
+            "Weybridge Rail Station",
+            "Weymouth Rail Station",
+            "South Wigston Rail Station",
+            "Wester Hailes Rail Station",
+            "Whalley Rail Station",
+            "Whaley Bridge Rail Station",
+            "West Ham Rail Station",
+            "Whitchurch (Cardiff) Rail Station",
+            "Woodhall Rail Station",
+            "White Hart Lane Rail Station",
+            "Whifflet Rail Station",
+            "Whimple Rail Station",
+            "Whinhill Rail Station",
+            "Whiston Rail Station",
+            "Whitehaven Rail Station",
+            "Whitland Rail Station",
+            "Whitlocks End Rail Station",
+            "West Hampstead Rail Station",
+            "West Hampstead Thameslink Rail Station",
+            "West Horndon Rail Station",
+            "Whitstable Rail Station",
+            "Whitley Bridge Rail Station",
+            "Whitchurch (Hants) Rail Station",
+            "Whittlesea Rail Station",
+            "Whitton Rail Station",
+            "Whitwell (Derbys) Rail Station",
+            "Whyteleafe Rail Station",
+            "Whyteleafe South Rail Station",
+            "Wick Rail Station",
+            "Widnes Rail Station",
+            "Wigan North Western Rail Station",
+            "Wigan Wallgate Rail Station",
+            "Wigton Rail Station",
+            "Wickford Rail Station",
+            "Wilmcote Rail Station",
+            "Wilnecote (Staffs) Rail Station",
+            "Wimbledon Rail Station",
+            "Wimbledon Chase Rail Station",
+            "Winchelsea Rail Station",
+            "Windsor & Eton Central Rail Station",
+            "Winnersh Rail Station",
+            "Winnersh Triangle Rail Station",
+            "Wishaw Rail Station",
+            "Witham (Essex) Rail Station",
+            "Witley Rail Station",
+            "Witton (West Midlands) Rail Station",
+            "Wivenhoe Rail Station",
+            "Wakefield Kirkgate Rail Station",
+            "Wakefield Westgate Rail Station",
+            "West Kilbride Rail Station",
+            "West Kirby Rail Station",
+            "Workington Rail Station",
+            "Wildmill Rail Station",
+            "Woldingham Rail Station",
+            "Wilmslow Rail Station",
+            "Wellingborough Rail Station",
+            "Willington Rail Station",
+            "London Waterloo East Rail Station",
+            "Waterloo (Merseyside) Rail Station",
+            "Willesden Junction Rail Station",
+            "Willesden Junction Low Level Rail Station",
+            "Walthamstow Queens Road Rail Station",
+            "Walthamstow Central Rail Station",
+            "Wolverton Rail Station",
+            "Welwyn Garden City Rail Station",
+            "Welwyn North Rail Station",
+            "West Malling Rail Station",
+            "Wembley Central Rail Station",
+            "Wembley Central Rail Station",
+            "Windermere Rail Station",
+            "Weston Milton Rail Station",
+            "Woodsmoor Rail Station",
+            "Williamwood Rail Station",
+            "Winchfield Rail Station",
+            "Winchester Rail Station",
+            "Wandsworth Road Rail Station",
+            "Winchmore Hill Rail Station",
+            "West Norwood Rail Station",
+            "White Notley Rail Station",
+            "Winsford Rail Station",
+            "Wanstead Park Rail Station",
+            "Woburn Sands Rail Station",
+            "Woodbridge Rail Station",
+            "Woking Rail Station",
+            "Wokingham Rail Station",
+            "Wolsingham Rail Station",
+            "Woolston Rail Station",
+            "Woolwich Dockyard Rail Station",
+            "Woolwich Arsenal Rail Station",
+            "Wombwell Rail Station",
+            "Walton-on-the-Naze Rail Station",
+            "Walton-on-Thames Rail Station",
+            "Woodley Rail Station",
+            "Wool Rail Station",
+            "Worcester Foregate Street Rail Station",
+            "Worcester Shrub Hill Rail Station",
+            "Worksop Rail Station",
+            "Worle Rail Station",
+            "Worstead Rail Station",
+            "Worthing Rail Station",
+            "Wootton Wawen Rail Station",
+            "Wrabness Rail Station",
+            "Warblington Rail Station",
+            "Worcester Park Rail Station",
+            "Wrenbury Rail Station",
+            "Wressle Rail Station",
+            "Wrexham Central Rail Station",
+            "Wrexham General Rail Station",
+            "Warrington Central Rail Station",
+            "Warrington Bank Quay Rail Station",
+            "Warminster Rail Station",
+            "Hoveton & Wroxham Rail Station",
+            "Worplesdon Rail Station",
+            "West Ruislip Rail Station",
+            "West Runton Rail Station",
+            "Wraysbury Rail Station",
+            "West St Leonards Rail Station",
+            "Weston-super-Mare Rail Station",
+            "Windsor & Eton Riverside Rail Station",
+            "Westbury (Wilts) Rail Station",
+            "Westhoughton Rail Station",
+            "Westerfield Rail Station",
+            "West Sutton Rail Station",
+            "Whitby Rail Station",
+            "Whitecraigs Rail Station",
+            "Whittlesford Parkway Rail Station",
+            "Wateringbury Rail Station",
+            "Water Orton Rail Station",
+            "Wivelsfield Rail Station",
+            "Wolverhampton Rail Station",
+            "Wavertree Technology Park Rail Station",
+            "West Wickham Rail Station",
+            "West Worthing Rail Station",
+            "Wye Rail Station",
+            "Wylde Green Rail Station",
+            "Wylam Rail Station",
+            "Wymondham Rail Station",
+            "Wythall Rail Station",
+            "Yarm Rail Station",
+            "Yalding Rail Station",
+            "Great Yarmouth Rail Station",
+            "Yate Rail Station",
+            "Yatton Rail Station",
+            "Yeoford Rail Station",
+            "Ynyswen Rail Station",
+            "Yoker Rail Station",
+            "York Rail Station",
+            "Yorton Rail Station",
+            "Yeovil Junction Rail Station",
+            "Yeovil Pen Mill Rail Station",
+            "Yardley Wood Rail Station",
+            "Ystrad Mynach Rail Station",
+            "Yetminster Rail Station",
+            "Ystrad Rhondda Rail Station",
+            "ZZZ",
+        };
+    }
+}
