+++ StationSearchCore.Web/Startup.cs
@@ -1,14 +1,10 @@
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.HttpsPolicy;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
+using StationSearchCore.Domain.Interfaces;
+using StationSearchCore.Service;
 namespace StationSearchCore.Web
 {
     public class Startup
@@ -24,6 +20,11 @@ namespace StationSearchCore.Web
         public void ConfigureServices(IServiceCollection services)
         {
             services.AddControllersWithViews();
+
+            // Should go inside an extention method on IServiceCollection called ~AddDomainServices()
+            services.AddTransient<ILookupService, LookupService>();
+            services.AddTransient<IPrefixTree, PrefixTree>();
+            services.AddTransient<IStationRepository, StationRepository>();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
