+++ StationSearchCore.Web/StationSearchCore.Web.csproj
@@ -1,10 +1,11 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\StationSearchCore.Domain\StationSearchCore.Domain.csproj" />
     <ProjectReference Include="..\StationSearchCore.Service\StationSearchCore.Service.csproj" />
   </ItemGroup>
